refactor(core): split Engine impl and re-check model checksums
Move the former engine.rs god-file into engine/{config,load,stream,transcribe,infer}.
Public Engine API is unchanged.

Before mapping weights, load now hashes every resolved file that has a pinned
digest. Production factories keep the check on; the mock factory skips it so
placeholder ONNX fixtures still load. Downloads also refuse a body larger than
2 GiB, advertised Content-Length included.
