--- crates/gigastt-core/src/inference/engine.rs
@@ -1,3889 +1,0 @@
-//! ONNX Runtime inference engine for GigaAM v3.
-
-use anyhow::Context;
-use std::path::Path;
-
-use crate::error::GigasttError;
-use crate::model::ModelVariant;
-#[allow(unused_imports)]
-use crate::runtime::factory::RuntimeFactory;
-use crate::runtime::production_factory_variant;
-use crate::runtime::tensor::{Shape, TensorDataView};
-
-use super::audio;
-use super::audio::{PcmWindows, SliceWindows, WindowSpec};
-use super::bias;
-use super::ctc;
-use super::decode;
-use super::load_files::{
-    ResolvedModelFiles, encoder_model_path, load_triplets_runtime, resolve_variant_required,
-};
-use super::pool::{Pool, SessionPool, SessionTriplet};
-use super::sizing;
-use super::state::{
-    DecoderState, EndpointMode, EndpointReason, FeatureExtractor, StreamingState,
-    TranscriptAssembler, TranscriptSegment, WordInfo, aggregate_confidence,
-};
-use super::token_format::{TokenFormatter, stitch_chunk_words};
-use super::tokenizer::Tokenizer;
-use super::types::{
-    DEFAULT_HOTWORDS_BOOST, DiarizationOutcome, HotwordError, HotwordOverride,
-    MAX_HOTWORD_PHRASE_CHARS, MAX_HOTWORDS_PER_REQUEST, OverrideError, TranscribeOverrides,
-    TranscribeRequest, TranscribeResult, TranscribeSource, merge_channel_results,
-};
-use super::windows::{
-    STREAM_DECODE_STRIDE_SAMPLES, STREAM_LEFT_CONTEXT_SAMPLES, STREAM_MAX_WINDOW_SAMPLES,
-    window_spec,
-};
-use super::{ENCODER_SUBSAMPLING, HOP_LENGTH, N_FFT, N_MELS, SECONDS_PER_FRAME, now_timestamp};
-
-#[cfg(feature = "diarization")]
-use super::diarization::{self, LazySpeakerEncoder};
-
-/// Cooperative-run hooks threaded through the decode call chain.
-///
-/// Both fields are `None` on the historical path, in which case every decode
-/// function behaves byte-for-byte as before — the hooks are only ever consulted
-/// through `if let Some(_)` guards, adding no work when absent. `abort` is
-/// polled at window boundaries so a cancelled run releases its pooled session
-/// within one window; `on_progress` receives the cumulative count of processed
-/// 16 kHz samples after each long-form window so a server watchdog can reset its
-/// no-progress deadline and drive a real progress bar.
-#[derive(Clone, Copy, Default)]
-struct DecodeControls<'a> {
-    abort: Option<&'a dyn Fn() -> bool>,
-    on_progress: Option<&'a dyn Fn(u64)>,
-}
-
-impl DecodeControls<'_> {
-    /// True once the caller has requested cancellation.
-    #[inline]
-    fn aborted(&self) -> bool {
-        self.abort.is_some_and(|a| a())
-    }
-
-    /// Report cumulative processed 16 kHz samples, if a sink is attached.
-    #[inline]
-    fn report(&self, processed_16k_samples: u64) {
-        if let Some(on_progress) = self.on_progress {
-            on_progress(processed_16k_samples);
-        }
-    }
-
-    /// Drop the progress sink but keep the abort hook. Used for the
-    /// `channels=split` path, where each channel restarts the sample clock and a
-    /// shared monotonic progress counter would go backwards.
-    #[inline]
-    fn abort_only(&self) -> Self {
-        Self {
-            abort: self.abort,
-            on_progress: None,
-        }
-    }
-}
-
-/// Default number of session triplets in the pool.
-///
-/// Each pooled triplet still materializes encoder weights; ORT's shared
-/// `PrepackedWeights` container (enabled on the CPU production factory) shares
-/// prepacked kernel buffers across sessions when the EP supports it — raw
-/// initializer tensors are not guaranteed to collapse to 1× (remeasure pool Δ
-/// after upgrades; see research T-002 / T-021). A pooled INT8 encoder triplet
-/// still costs on the order of ~0.4 GB resident in practice, so the default
-/// pool stays at 2. Raise `--pool-size` when higher concurrency is needed and
-/// RAM allows.
-#[cfg(target_os = "android")]
-const DEFAULT_POOL_SIZE: usize = 1;
-#[cfg(not(target_os = "android"))]
-const DEFAULT_POOL_SIZE: usize = 2;
-
-/// ONNX Runtime inference engine for GigaAM v3 e2e_rnnt.
-///
-/// Thread-safe: inference sessions live in a [`SessionPool`] so `Engine` can be
-/// shared across connections via `Arc<Engine>`. The pool size acts as the
-/// concurrency limit — no separate semaphore needed. Typical usage:
-///
-/// ```ignore
-/// let engine = Engine::load("~/.gigastt/models")?;
-/// let mut guard = engine.pool.checkout().await?;
-/// let text = engine.transcribe_file("audio.wav", &mut guard)?;
-/// // guard is returned to the pool on drop
-/// ```
-///
-/// For streaming recognition, use [`create_state`](Engine::create_state) +
-/// [`process_chunk`](Engine::process_chunk) + [`flush_state`](Engine::flush_state).
-pub struct Engine {
-    /// Pool of session triplets for interactive inference (WebSocket + SSE
-    /// streaming). REST file transcription uses [`Engine::batch_pool`] when it
-    /// is set, so a long batch job can't starve real-time streaming.
-    pub pool: SessionPool,
-    /// Optional dedicated pool for batch REST file transcription, split off
-    /// from `pool` at load time. `None` means REST shares the interactive pool.
-    pub batch_pool: Option<SessionPool>,
-    tokenizer: Tokenizer,
-    features: FeatureExtractor,
-    /// Recognition head detected on disk at load time. Drives the default
-    /// punctuation policy (`auto`): on for [`ModelVariant::Rnnt`] (bare output),
-    /// off for [`ModelVariant::E2eRnnt`] (already punctuated).
-    variant: ModelVariant,
-    /// Optional punctuation / casing restorer applied to file-transcription
-    /// output and to finalized streaming segments. `None` = pass-through (the
-    /// default, and the only behaviour when no punct model is installed).
-    /// Attached via [`Engine::with_punctuator`].
-    punctuator: Option<crate::punctuation::Punctuator>,
-    /// Whether to run inverse text normalization (Russian number-words →
-    /// digits) on file-transcription output and finalized streaming segments,
-    /// *before* the punctuation pass. Off by default; toggled via
-    /// [`Engine::with_itn`].
-    itn: bool,
-    /// Optional contextual hotword biaser applied inside the greedy RNN-T decode
-    /// loop (shallow fusion). `None` = no biasing (the default), and the decode
-    /// path is then byte-for-byte identical to the un-biased engine. Attached
-    /// via [`Engine::with_biaser`]. Shared across the session pool by reference.
-    biaser: Option<bias::Biaser>,
-    /// Optional Silero VAD. When set, file transcription skips silent regions
-    /// (decoding only detected speech) and streaming endpointing is owned by
-    /// VAD-detected trailing silence (the decoder's blank-run heuristic is
-    /// ignored, so `vad_config.min_silence_ms` fully controls finalization).
-    /// `None` = no VAD: the file path decodes the whole buffer and streaming
-    /// endpointing is byte-for-byte unchanged. Attached via [`Engine::with_vad`].
-    vad: Option<crate::vad::SileroVad>,
-    /// Thresholds for the VAD (speech threshold, min silence/speech, padding).
-    /// Ignored when `vad` is `None`.
-    vad_config: crate::vad::VadConfig,
-    /// Default streaming utterance-end policy for new sessions. Overridable
-    /// per session via WS `configure.endpoint_mode`.
-    endpoint_mode: EndpointMode,
-    /// Whether the INT8 quantized encoder is in use.
-    int8: bool,
-    /// True when pooled encoder sessions run on the ANE fixed-shape pad-up path.
-    /// Selects the 30s long-form chunk window (vs 24s for ort). Derived from
-    /// the loaded encoder session at boot, not from compile-time features alone,
-    /// so non-rnnt heads on an `ane`-feature binary still use the ort window.
-    ane_encoder: bool,
-    /// Lazy speaker encoder for diarization (`None` if model file is absent).
-    ///
-    /// Boot only probes for `wespeaker_resnet34.onnx`; the ONNX session is
-    /// opened on the first diarization request so unused speaker files do not
-    /// inflate ready RSS (~+39 MiB when loaded). Shared across sessions via
-    /// the `Arc` inside the loaded encoder.
-    #[cfg(feature = "diarization")]
-    speaker_encoder: Option<LazySpeakerEncoder>,
-}
-
-impl Engine {
-    /// Whether the INT8 quantized encoder is loaded.
-    pub fn is_int8(&self) -> bool {
-        self.int8
-    }
-
-    /// The recognition head ([`ModelVariant`]) detected on disk at load time.
-    /// Lets callers decide the default punctuation policy (`auto`).
-    pub fn variant(&self) -> ModelVariant {
-        self.variant
-    }
-
-    /// Attach an optional punctuation / casing restorer, consuming and
-    /// returning `self` (builder style). Pass `None` for pass-through. When set,
-    /// the restorer post-processes the final text of file transcription
-    /// ([`Engine::transcribe_file`] / [`Engine::transcribe_bytes_shared`]) and
-    /// of finalized streaming segments ([`Engine::process_chunk`] /
-    /// [`Engine::flush_state`]).
-    pub fn with_punctuator(mut self, punctuator: Option<crate::punctuation::Punctuator>) -> Self {
-        self.punctuator = punctuator;
-        self
-    }
-
-    /// Whether a punctuation restorer is attached.
-    pub fn has_punctuator(&self) -> bool {
-        self.punctuator.is_some()
-    }
-
-    /// Reject a [`TranscribeOverrides`] that turns a knob on without the backing
-    /// resource loaded. Call this *before* checking out a session so the request
-    /// fails fast (the REST layer maps the error to a `409`). Turning a knob off
-    /// (`Some(false)`) and ITN in either direction are always valid — ITN is pure
-    /// code with no model to load.
-    ///
-    /// Hotword DoS limits are validated separately via
-    /// [`Engine::validate_hotwords`] so [`TranscribeOverrides`] stays `Copy` (semver).
-    ///
-    /// # Errors
-    ///
-    /// - [`OverrideError::VadNotLoaded`] when `vad = Some(true)` but no VAD is
-    ///   attached.
-    /// - [`OverrideError::PunctuationNotAvailable`] when `punctuation = Some(true)`
-    ///   but no punctuator is attached.
-    pub fn validate_overrides(&self, o: &TranscribeOverrides) -> Result<(), OverrideError> {
-        if o.vad == Some(true) && self.vad.is_none() {
-            return Err(OverrideError::VadNotLoaded);
-        }
-        if o.punctuation == Some(true) && self.punctuator.is_none() {
-            return Err(OverrideError::PunctuationNotAvailable);
-        }
-        Ok(())
-    }
-
-    /// Reject a [`HotwordOverride`] that exceeds DoS limits. Call before checkout
-    /// so oversized requests fail fast (REST maps to HTTP 400).
-    ///
-    /// # Errors
-    ///
-    /// - [`HotwordError::TooManyHotwords`] when more than
-    ///   [`MAX_HOTWORDS_PER_REQUEST`] phrases are supplied.
-    /// - [`HotwordError::PhraseTooLong`] when any phrase exceeds
-    ///   [`MAX_HOTWORD_PHRASE_CHARS`] Unicode scalar values.
-    pub fn validate_hotwords(&self, hw: &HotwordOverride) -> Result<(), HotwordError> {
-        if hw.phrases.len() > MAX_HOTWORDS_PER_REQUEST {
-            return Err(HotwordError::TooManyHotwords);
-        }
-        for phrase in &hw.phrases {
-            if phrase.chars().count() > MAX_HOTWORD_PHRASE_CHARS {
-                return Err(HotwordError::PhraseTooLong);
-            }
-        }
-        Ok(())
-    }
-
-    /// Build a temporary per-request [`bias::Biaser`] from a
-    /// [`HotwordOverride`]. Empty phrases force biasing off (`None`).
-    /// Unrepresentable phrases are dropped by [`Biaser::from_phrases`]; if none
-    /// survive, returns `None` (decode continues without biasing).
-    fn build_request_biaser(&self, hw: &HotwordOverride) -> Option<bias::Biaser> {
-        if hw.phrases.is_empty() {
-            return None;
-        }
-        let boost = hw.boost.unwrap_or(DEFAULT_HOTWORDS_BOOST);
-        let pairs: Vec<(String, f32)> = hw.phrases.iter().cloned().map(|p| (p, 1.0)).collect();
-        bias::Biaser::from_phrases(&self.tokenizer, &pairs, boost)
-    }
-
-    /// Effective hotword biaser for one decode call.
-    ///
-    /// - `hotwords == None` → engine boot biaser (may itself be `None`)
-    /// - `hotwords == Some(_)` → temporary request biaser (may be `None` when
-    ///   phrases are empty / unrepresentable — deliberately *not* the boot biaser)
-    ///
-    /// `request_biaser` must be the result of [`build_request_biaser`] (or
-    /// `None` when `hotwords` is absent); it is only borrowed, never moved.
-    fn select_biaser<'a>(
-        &'a self,
-        hotwords: Option<&HotwordOverride>,
-        request_biaser: &'a Option<bias::Biaser>,
-    ) -> Option<&'a bias::Biaser> {
-        match hotwords {
-            None => self.biaser.as_ref(),
-            Some(_) => request_biaser.as_ref(),
-        }
-    }
-
-    /// Apply ITN then punctuation to joined text. Shared by the streaming
-    /// finalizer and the file-transcription result builder so the order and
-    /// guards stay identical.
-    fn apply_text_postprocess(&self, text: String, itn: bool, punctuation: bool) -> String {
-        let text = if itn {
-            crate::itn::apply_itn(&text)
-        } else {
-            text
-        };
-        match &self.punctuator {
-            Some(p) if punctuation => p.restore(&text),
-            _ => text,
-        }
-    }
-
-    /// Enable or disable inverse text normalization (Russian number-words →
-    /// digits) on file-transcription output and finalized streaming segments,
-    /// consuming and returning `self` (builder style). When enabled, ITN runs
-    /// *before* the punctuation pass so the restorer cases the
-    /// already-digitized text.
-    pub fn with_itn(mut self, enabled: bool) -> Self {
-        self.itn = enabled;
-        self
-    }
-
-    /// Whether inverse text normalization is enabled.
-    pub fn has_itn(&self) -> bool {
-        self.itn
-    }
-
-    /// Attach a contextual hotword biaser built from `(phrase, weight)` pairs
-    /// and an additive `boost`, consuming and returning `self` (builder style).
-    /// Each phrase is tokenized with the engine's own `Tokenizer`, so biasing
-    /// adapts to whichever recognition head is loaded.
-    ///
-    /// When `phrases` is empty, `boost <= 0`, or no phrase is representable in
-    /// the active vocab, the biaser resolves to `None` and the decode path stays
-    /// byte-for-byte unchanged. Replaces any previously attached biaser.
-    pub fn with_hotwords(mut self, phrases: &[(String, f32)], boost: f32) -> Self {
-        self.biaser = if phrases.is_empty() {
-            None
-        } else {
-            bias::Biaser::from_phrases(&self.tokenizer, phrases, boost)
-        };
-        if let Some(b) = &self.biaser {
-            tracing::info!(
-                "Hotword biasing enabled ({} phrase(s), boost {boost})",
-                b.phrase_count()
-            );
-        }
-        self
-    }
-
-    /// Whether a hotword biaser is attached (biasing active).
-    pub fn has_hotwords(&self) -> bool {
-        self.biaser.is_some()
-    }
-
-    /// Attach an optional Silero VAD plus its config, consuming and returning
-    /// `self` (builder style). Pass `None` for no VAD (the default): file
-    /// transcription then decodes the whole buffer and streaming endpointing is
-    /// byte-for-byte unchanged. When set, file transcription skips silence and
-    /// streaming endpointing is owned by VAD-detected trailing silence (the
-    /// decoder's blank-run heuristic is ignored).
-    pub fn with_vad(
-        mut self,
-        vad: Option<crate::vad::SileroVad>,
-        config: crate::vad::VadConfig,
-    ) -> Self {
-        self.vad = vad;
-        self.vad_config = config;
-        if self.vad.is_some() {
-            tracing::info!(
-                "VAD enabled (threshold {}, min_silence {}ms)",
-                self.vad_config.threshold,
-                self.vad_config.min_silence_ms
-            );
-        }
-        self
-    }
-
-    /// Whether a VAD is attached (silence skipping / VAD endpointing active).
-    pub fn has_vad(&self) -> bool {
-        self.vad.is_some()
-    }
-
-    /// Boot-time VAD config (threshold, min silence, …). Useful when recreating
-    /// a per-session endpointer after `configure.min_silence_ms`.
-    pub fn vad_config(&self) -> &crate::vad::VadConfig {
-        &self.vad_config
-    }
-
-    /// Boot-time streaming endpoint mode for new sessions.
-    pub fn endpoint_mode(&self) -> EndpointMode {
-        self.endpoint_mode
-    }
-
-    /// Set the default streaming utterance-end policy for new sessions.
-    pub fn with_endpoint_mode(mut self, mode: EndpointMode) -> Self {
-        self.endpoint_mode = mode;
-        self
-    }
-
-    /// Size of the BPE vocabulary the loaded tokenizer covers. Exposed so the
-    /// REST `/v1/models` handler can report the real value instead of a
-    /// hardcoded literal that would drift if the upstream model rev changes.
-    pub fn vocab_size(&self) -> usize {
-        self.tokenizer.vocab_size()
-    }
-
-    /// Load ONNX models from the given directory and create an inference engine.
-    ///
-    /// Creates a pool of `DEFAULT_POOL_SIZE` session triplets for concurrent inference.
-    /// The recognition head ([`ModelVariant`]) is auto-detected from the encoder
-    /// file present on disk: `v3_rnnt_encoder.onnx` (or `_int8.onnx`) selects the
-    /// plain rnnt head, else `v3_e2e_rnnt_encoder.onnx` selects e2e_rnnt. The
-    /// matching decoder, joiner, and vocab files must also be present.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`GigasttError::ModelLoad`] if model files are missing or ONNX session creation fails.
-    pub fn load(model_dir: &str) -> Result<Self, GigasttError> {
-        Self::load_with_pool_size(model_dir, DEFAULT_POOL_SIZE)
-    }
-
-    /// Load ONNX models with a custom pool size. Requires the *full* pool to
-    /// load (every triplet); use [`Engine::load_with_pool_size_min`] to boot on
-    /// a partial pool.
-    pub fn load_with_pool_size(model_dir: &str, pool_size: usize) -> Result<Self, GigasttError> {
-        Self::load_with_pool_size_min(model_dir, pool_size, pool_size)
-    }
-
-    /// Load ONNX models with a custom pool size, tolerating a partial pool down
-    /// to `min_size` triplets when the rest fail to load. Boots a degraded pool
-    /// with a warning when `min_size <= loaded < pool_size`; errors only when
-    /// fewer than `min_size` triplets load. `min_size` is clamped to
-    /// `1..=pool_size`.
-    pub fn load_with_pool_size_min(
-        model_dir: &str,
-        pool_size: usize,
-        min_size: usize,
-    ) -> Result<Self, GigasttError> {
-        // No batch/stream split: the whole pool is interactive.
-        Self::load_with_pools(model_dir, pool_size, min_size, 0)
-    }
-
-    /// Load ONNX models splitting the pool into an interactive pool (WebSocket +
-    /// SSE) and a dedicated batch pool of `batch_pool_size` triplets for REST
-    /// file transcription, so a long batch job can't starve real-time
-    /// streaming. `batch_pool_size == 0` disables the split (REST shares the
-    /// interactive pool); it is clamped to leave at least one interactive
-    /// triplet. Partial-load tolerance follows `min_size` as in
-    /// [`Engine::load_with_pool_size_min`].
-    pub fn load_with_pools(
-        model_dir: &str,
-        pool_size: usize,
-        min_size: usize,
-        batch_pool_size: usize,
-    ) -> Result<Self, GigasttError> {
-        Self::load_with_pools_threads(model_dir, pool_size, min_size, batch_pool_size, 1)
-    }
-
-    /// Like [`Engine::load_with_pools`], but with a configurable encoder
-    /// intra-op thread count for the CPU EP. `encoder_intra_threads == 1` (the
-    /// default everywhere else) builds sessions identical to the prior
-    /// behaviour. Values `> 1` give the dominant encoder more intra-op
-    /// parallelism on weak CPUs / long single-file jobs; the count is clamped
-    /// against the logical CPU count so `pool_size * threads` can't oversubscribe
-    /// the machine (see `Engine::clamp_encoder_intra_threads`). Ignored by the
-    /// CoreML / CUDA builds (the accelerator owns scheduling there).
-    pub fn load_with_pools_threads(
-        model_dir: &str,
-        pool_size: usize,
-        min_size: usize,
-        batch_pool_size: usize,
-        encoder_intra_threads: usize,
-    ) -> Result<Self, GigasttError> {
-        Self::load_with_pools_threads_variant(
-            model_dir,
-            None,
-            pool_size,
-            min_size,
-            batch_pool_size,
-            encoder_intra_threads,
-        )
-    }
-
-    /// Like [`Engine::load_with_pools_threads`], but lets the caller force which
-    /// recognition head to load instead of auto-detecting it.
-    ///
-    /// When `variant` is `Some(v)`, head `v` is loaded (and the load fails with a
-    /// clear `ModelLoad` error if `v`'s files aren't in `model_dir`). When it is
-    /// `None`, the on-disk layout is auto-detected with `rnnt` precedence, exactly
-    /// as [`Engine::load_with_pools_threads`]. This is the entry point that makes
-    /// `--model-variant` effective when a directory holds more than one head.
-    pub fn load_with_pools_threads_variant(
-        model_dir: &str,
-        variant: Option<ModelVariant>,
-        pool_size: usize,
-        min_size: usize,
-        batch_pool_size: usize,
-        encoder_intra_threads: usize,
-    ) -> Result<Self, GigasttError> {
-        let dir = Path::new(model_dir);
-        // Resolve the head once, up front: an explicit `variant` wins, else
-        // manifest.toml architecture, else detect from disk (rnnt precedence).
-        // Resolving here (not just inside `load_with_factory`) keeps the RAM cap
-        // sized to the head that will actually load.
-        let variant = resolve_variant_required(variant, dir)?;
-        // Bound the idle footprint: each pooled triplet deserializes its own
-        // encoder copy, so a large `--pool-size` on a small host can OOM at
-        // load. Clamp by available RAM (logs when it clamps); a no-op on hosts
-        // with ample memory.
-        let encoder_bytes = std::fs::metadata(encoder_model_path(dir, variant))
-            .map(|m| m.len())
-            .unwrap_or(0);
-        let pool_size =
-            Self::cap_pool_size_for_ram(pool_size, encoder_bytes, sizing::effective_ram_bytes());
-        // Don't let `pool_size * encoder_intra_threads` oversubscribe the CPU
-        // (no-op when the default `1` is requested).
-        let logical_cpus = std::thread::available_parallelism()
-            .map(|n| n.get())
-            .unwrap_or(1);
-        let encoder_intra_threads =
-            Self::clamp_encoder_intra_threads(pool_size, encoder_intra_threads, logical_cpus);
-
-        let factory = production_factory_variant(dir, Some(variant));
-        Self::load_with_factory(
-            dir,
-            Some(variant),
-            pool_size,
-            min_size,
-            batch_pool_size,
-            factory,
-            encoder_intra_threads,
-        )
-    }
-
-    /// Assemble an engine from already-loaded sessions, tokenizer, and variant.
-    ///
-    /// Optional attachments (punctuator, ITN, biaser, VAD) stay off — callers
-    /// chain [`Engine::with_punctuator`] / [`Engine::with_itn`] / etc.
-    fn from_loaded_parts(
-        pool: SessionPool,
-        batch_pool: Option<SessionPool>,
-        tokenizer: Tokenizer,
-        variant: ModelVariant,
-        int8: bool,
-        ane_encoder: bool,
-        #[cfg(feature = "diarization")] speaker_encoder: Option<LazySpeakerEncoder>,
-    ) -> Self {
-        Self {
-            pool,
-            batch_pool,
-            tokenizer,
-            features: FeatureExtractor::new(),
-            variant,
-            punctuator: None,
-            itn: false,
-            biaser: None,
-            vad: None,
-            vad_config: crate::vad::VadConfig::default(),
-            endpoint_mode: EndpointMode::Auto,
-            int8,
-            ane_encoder,
-            #[cfg(feature = "diarization")]
-            speaker_encoder,
-        }
-    }
-
-    /// Package-private factory-based loader. Used by production code paths and
-    /// by tests that inject a [`crate::runtime::factory::RuntimeFactory`].
-    pub(crate) fn load_with_factory(
-        model_dir: &Path,
-        variant_override: Option<ModelVariant>,
-        pool_size: usize,
-        min_size: usize,
-        batch_pool_size: usize,
-        factory: Box<dyn crate::runtime::factory::RuntimeFactory>,
-        encoder_intra_threads: usize,
-    ) -> Result<Self, GigasttError> {
-        // Honor an explicit variant (e.g. from `--model-variant`) when the caller
-        // resolved one; otherwise prefer `manifest.toml`, else auto-detect from
-        // the on-disk layout (rnnt precedence). Passing the override through is
-        // what makes `--model-variant` effective when a directory holds more than
-        // one head — without it the engine always re-detects and silently loads
-        // the highest-precedence head.
-        let variant = resolve_variant_required(variant_override, model_dir)?;
-        let pool_size = pool_size.max(1);
-        let min_size = min_size.clamp(1, pool_size);
-
-        let runtime = factory.create(encoder_intra_threads)?;
-        let model_load = |e: anyhow::Error| GigasttError::ModelLoad {
-            path: model_dir.display().to_string(),
-            source: Some(e.into()),
-        };
-
-        let files = ResolvedModelFiles::resolve(model_dir, variant).map_err(model_load)?;
-
-        tracing::info!("Detected model variant: {variant:?}");
-        // The candle backend loads FP32 `candle/*.safetensors` and ignores the
-        // INT8 ONNX encoder, so report no INT8 there and gate out the INT8 / ONNX
-        // / CPU-EP logs below (which are wrong for candle), emitting an accurate
-        // candle line instead. The default (ort) logging is unchanged.
-        let is_int8 = !cfg!(feature = "candle") && files.using_int8;
-        if !cfg!(feature = "candle") {
-            // ORT product path is INT8-only (`ResolvedModelFiles` rejects FP32).
-            tracing::info!("Using INT8 quantized encoder");
-            tracing::info!(
-                "Loading ONNX models from {} (pool_size={pool_size})...",
-                model_dir.display()
-            );
-        }
-
-        #[cfg(feature = "candle")]
-        tracing::info!(
-            "Loading Candle models from {} (pool_size={pool_size})...",
-            model_dir.display()
-        );
-        #[cfg(feature = "coreml")]
-        tracing::info!("Using CoreML execution provider (Neural Engine + CPU)");
-        #[cfg(feature = "cuda")]
-        tracing::info!("Using CUDA execution provider (falls back to CPU if unavailable)");
-        #[cfg(not(any(feature = "coreml", feature = "cuda", feature = "candle")))]
-        tracing::info!("Using CPU execution provider");
-
-        // CoreML can reject a model at load time; fall back to CPU if that happens.
-        #[cfg(feature = "coreml")]
-        let triplets = match load_triplets_runtime(&*runtime, &files, variant, pool_size, min_size)
-            .map_err(model_load)
-        {
-            Ok(triplets) => triplets,
-            Err(load_err) => {
-                tracing::warn!(
-                    "CoreML EP failed to load sessions ({load_err:#}); falling back to CPU execution provider"
-                );
-                let cpu_factory = factory.cpu_fallback();
-                let runtime = cpu_factory.create(encoder_intra_threads)?;
-                load_triplets_runtime(&*runtime, &files, variant, pool_size, min_size)
-                    .map_err(model_load)?
-            }
-        };
-        #[cfg(not(feature = "coreml"))]
-        let triplets = load_triplets_runtime(&*runtime, &files, variant, pool_size, min_size)
-            .map_err(model_load)?;
-
-        let tokenizer = Tokenizer::load(&files.vocab).map_err(model_load)?;
-
-        tracing::info!(
-            "Models loaded (vocab_size={}, pool_size={pool_size})",
-            tokenizer.vocab_size()
-        );
-
-        // Probe only — do not open the WeSpeaker ONNX session at boot. The
-        // encoder is loaded on the first diarization request.
-        #[cfg(feature = "diarization")]
-        let speaker_encoder = diarization::probe_speaker_encoder(model_dir);
-
-        // Detect ANE from the loaded encoder session (not compile-time alone) so
-        // non-rnnt heads / injected factories keep the ort chunk window.
-        let ane_encoder = triplets.first().is_some_and(|t| t.encoder.is_ane_encoder());
-        let (pool, batch_pool) = Self::split_triplets(triplets, batch_pool_size);
-        let engine = Self::from_loaded_parts(
-            pool,
-            batch_pool,
-            tokenizer,
-            variant,
-            is_int8,
-            ane_encoder,
-            #[cfg(feature = "diarization")]
-            speaker_encoder,
-        );
-
-        // CoreML compiles its graph partitions lazily, so sessions that loaded
-        // fine can still fail at the first `Run()`. Probe one triplet now; if the
-        // probe fails, rebuild the pool on the CPU EP.
-        #[cfg(feature = "coreml")]
-        let engine = sizing::probe_or_rebuild(
-            engine,
-            |e: &Self| e.warmup_one().map_err(anyhow::Error::from),
-            |mut e, probe_err| {
-                tracing::warn!(
-                    "CoreML EP failed at runtime ({probe_err:#}); falling back to CPU execution provider"
-                );
-                let cpu_factory = factory.cpu_fallback();
-                let runtime = cpu_factory
-                    .create(encoder_intra_threads)
-                    .map_err(|e| anyhow::anyhow!(e))?;
-                let triplets =
-                    load_triplets_runtime(&*runtime, &files, variant, pool_size, min_size)?;
-                let (pool, batch_pool) = Self::split_triplets(triplets, batch_pool_size);
-                e.pool = pool;
-                e.batch_pool = batch_pool;
-                Ok(e)
-            },
-        )
-        .map_err(model_load)?;
-
-        Ok(engine)
-    }
-
-    /// Split loaded triplets into an interactive pool and an optional batch
-    /// pool of `batch_pool_size` triplets. Always leaves at least one triplet
-    /// for the interactive pool; `batch_pool_size == 0` (or a pool too small to
-    /// split) yields no batch pool.
-    fn split_triplets(
-        triplets: Vec<SessionTriplet>,
-        batch_pool_size: usize,
-    ) -> (SessionPool, Option<SessionPool>) {
-        Self::split_pool(triplets, batch_pool_size)
-    }
-
-    /// Generic pool split underlying [`Engine::split_triplets`]: partition
-    /// `items` into an interactive pool and an optional batch pool of
-    /// `batch_pool_size` items, always leaving at least one item interactive.
-    /// `batch_pool_size == 0` (or too few items to split) yields no batch pool.
-    /// Generic over the item type so the routing can be unit-tested with a
-    /// synthetic `Pool<u32>` instead of model-backed `SessionTriplet`s.
-    fn split_pool<T: Send>(items: Vec<T>, batch_pool_size: usize) -> (Pool<T>, Option<Pool<T>>) {
-        let (interactive, batch) = sizing::split_pool_items(items, batch_pool_size);
-        (Pool::new(interactive), batch.map(Pool::new))
-    }
-
-    /// See [`sizing::cap_pool_size_for_ram`].
-    fn cap_pool_size_for_ram(requested: usize, encoder_bytes: u64, total_ram: u64) -> usize {
-        sizing::cap_pool_size_for_ram(requested, encoder_bytes, total_ram)
-    }
-
-    /// See [`sizing::clamp_encoder_intra_threads`].
-    fn clamp_encoder_intra_threads(
-        pool_size: usize,
-        requested: usize,
-        logical_cpus: usize,
-    ) -> usize {
-        sizing::clamp_encoder_intra_threads(pool_size, requested, logical_cpus)
-    }
-
-    /// Run one ~1 s silent inference on a single pooled session triplet.
-    ///
-    /// Exercises the full mel + encoder + RNN-T decode pipeline, forcing
-    /// lazy EP work (CoreML partition compilation, first-run allocations) to
-    /// happen now instead of on the first real request — and doubling as a
-    /// runtime self-check for EPs that can fail at prediction time even
-    /// though their sessions loaded fine (issue #42).
-    fn warmup_one(&self) -> Result<(), GigasttError> {
-        self.warmup_one_on(&self.pool)
-    }
-
-    /// Warm a single triplet from a specific pool with ~1 s of silence.
-    fn warmup_one_on(&self, pool: &SessionPool) -> Result<(), GigasttError> {
-        let silence = vec![0.0f32; 16000]; // 1 s at 16 kHz
-        let mut guard = pool
-            .checkout_blocking()
-            .map_err(|e| GigasttError::Inference {
-                source: Box::new(e),
-            })?;
-        self.transcribe_samples(&silence, &mut guard)?;
-        Ok(())
-    }
-
-    /// Warm up every pooled session triplet with a ~1 s silent inference
-    /// so the first real request doesn't pay the EP compile /
-    /// first-allocation cost.
-    ///
-    /// Sequential checkouts visit each pooled triplet exactly once because
-    /// check-in returns items to the back of the FIFO queue.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`GigasttError::Inference`] if a warmup inference fails — with
-    /// the `coreml` feature this is unexpected, because [`Engine::load`]
-    /// already probed the pool and fell back to the CPU EP if needed.
-    pub fn warmup(&self) -> Result<(), GigasttError> {
-        for _ in 0..self.pool.total() {
-            self.warmup_one()?;
-        }
-        if let Some(ref batch) = self.batch_pool {
-            for _ in 0..batch.total() {
-                self.warmup_one_on(batch)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// The pool REST file transcription should use: the dedicated batch pool
-    /// when one was split off, otherwise the interactive pool.
-    pub fn pool_for_batch(&self) -> &SessionPool {
-        self.batch_pool.as_ref().unwrap_or(&self.pool)
-    }
-
-    /// Close both the interactive and batch pools so every waiter wakes with
-    /// `PoolError::Closed` during graceful shutdown.
-    pub fn close_pools(&self) {
-        self.pool.close();
-        if let Some(ref batch) = self.batch_pool {
-            batch.close();
-        }
-    }
-
-    /// Return `true` if a speaker model file was present at boot and diarization
-    /// can be requested. The ONNX session may still be unloaded until the first
-    /// diarization request (lazy load).
-    #[cfg(feature = "diarization")]
-    pub fn has_speaker_encoder(&self) -> bool {
-        self.speaker_encoder.is_some()
-    }
-
-    /// Create a fresh streaming state for a new connection.
-    ///
-    /// Pass `diarization_enabled = true` to activate speaker diarization for
-    /// this session. Without the `diarization` feature or a speaker model file,
-    /// the flag is silently ignored (a `warn!` is emitted when the caller asked
-    /// for diarization but the build does not support it, so the contract
-    /// mismatch is visible in logs). Enabling diarization loads the speaker
-    /// encoder on first use if it was only probed at boot.
-    pub fn create_state(&self, diarization_enabled: bool) -> StreamingState {
-        #[cfg(feature = "diarization")]
-        let diarization_state = if diarization_enabled {
-            match self.speaker_encoder.as_ref() {
-                Some(lazy) => lazy
-                    .get_or_load()
-                    .and_then(|enc| diarization::open_streaming(&enc)),
-                None => {
-                    tracing::warn!(
-                        "diarization_enabled=true ignored: wespeaker model not present at engine boot"
-                    );
-                    None
-                }
-            }
-        } else {
-            None
-        };
-
-        #[cfg(not(feature = "diarization"))]
-        if diarization_enabled {
-            tracing::warn!(
-                "diarization_enabled=true ignored: build lacks the `diarization` feature"
-            );
-        }
-
-        StreamingState {
-            decoder: DecoderState::new(self.tokenizer.blank_id()),
-            audio_buffer: Vec::new(),
-            assembler: TranscriptAssembler::new(),
-            window_start_samples: 0,
-            context_samples: 0,
-            pending_samples: 0,
-            resampler: None,
-            mel_fft_input: Vec::new(),
-            mel_power: Vec::new(),
-            mel_output: Vec::new(),
-            resample_output_buf: Vec::new(),
-            vad_endpointer: self
-                .vad
-                .as_ref()
-                .map(|_| crate::vad::VadEndpointer::new(&self.vad_config)),
-            punctuation: None,
-            itn: None,
-            endpoint_mode: self.endpoint_mode,
-            #[cfg(feature = "diarization")]
-            diarization_state,
-        }
-    }
-
-    /// Process a chunk of 16kHz f32 audio samples and return any new transcript segments.
-    ///
-    /// Returns [`TranscriptSegment`] with `is_final == false` during speech (Partial),
-    /// and `is_final == true` only on a **true utterance endpoint**:
-    ///
-    /// - decoder blank-run (~600 ms) when no VAD and [`EndpointMode::Auto`];
-    /// - VAD trailing silence when a VAD is attached (and mode is not `Manual`);
-    /// - never on the ~2.5 s encoder window cap (cap commits a stable prefix and
-    ///   emits a non-final partial so voice assistants do not treat a slide as
-    ///   "command complete").
-    ///
-    /// Streaming state (LSTM hidden/cell, leftover audio, accumulated text) is
-    /// maintained in `state`.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`GigasttError::Inference`] if the ONNX runtime fails.
-    pub fn process_chunk(
-        &self,
-        samples: &[f32],
-        state: &mut StreamingState,
-        triplet: &mut SessionTriplet,
-    ) -> Result<Vec<TranscriptSegment>, GigasttError> {
-        if samples.is_empty() {
-            return Ok(vec![]);
-        }
-
-        // Diarization tracks speakers continuously, so feed every chunk's audio
-        // even when this chunk doesn't trigger a decode (see the stride gate).
-        #[cfg(feature = "diarization")]
-        if let Some(dia) = state.diarization_state.as_mut() {
-            diarization::feed_chunk(dia, samples);
-        }
-
-        // Sliding-window streaming: accumulate audio; the encoder re-runs on the
-        // whole retained window so the offline Conformer always has left context
-        // (an isolated ~100ms chunk decodes to garbage). Re-decoding is the cost,
-        // so we only decode once STREAM_DECODE_STRIDE_SAMPLES of NEW audio have
-        // arrived (or the window hit its cap) — this keeps the engine real-time.
-        // The window is bounded by STREAM_MAX_WINDOW_SAMPLES; on endpoint or cap
-        // we finalize the tail and slide, retaining STREAM_LEFT_CONTEXT_SAMPLES.
-        state.audio_buffer.extend_from_slice(samples);
-        state.pending_samples += samples.len();
-
-        // Feed the VAD on every chunk so trailing silence is tracked
-        // continuously (independent of the decode stride). A VAD endpoint forces
-        // a decode + finalize this chunk even if the stride gate wouldn't fire.
-        // VAD is non-blocking: an inference error is logged and ignored, leaving
-        // the window cap as the only backstop until the VAD recovers. With no
-        // VAD attached `vad_endpoint` is always false and the decoder's
-        // blank-run heuristic keeps owning endpointing, byte-for-byte unchanged.
-        let mut vad_endpoint = false;
-        if let (Some(vad), Some(ep)) = (self.vad.as_ref(), state.vad_endpointer.as_mut()) {
-            match ep.push(vad, samples) {
-                Ok(fired) => vad_endpoint = fired,
-                Err(e) => tracing::warn!("VAD endpoint detection failed: {e:#}"),
-            }
-        }
-
-        let over_cap = state.audio_buffer.len() >= STREAM_MAX_WINDOW_SAMPLES;
-        // Stride gate on NEW audio since the last decode (not since the last
-        // slide): otherwise a non-finalizing partial would leave the counter
-        // high and decode on every subsequent chunk. A VAD endpoint overrides
-        // the gate so the utterance finalizes promptly.
-        if state.pending_samples < STREAM_DECODE_STRIDE_SAMPLES && !over_cap && !vad_endpoint {
-            return Ok(vec![]);
-        }
-        // Too little audio to extract a frame. Skip — but never when finalizing:
-        // a fired VAD endpoint (or cap) must still flush the assembler below,
-        // even though `decode_window` will add no new words from a sub-frame
-        // buffer. (In practice a VAD endpoint needs ≥ min_silence_ms of trailing
-        // audio, so the buffer is always ≫ N_FFT here; this guards the edge.)
-        if state.audio_buffer.len() < N_FFT && !vad_endpoint && !over_cap {
-            return Ok(vec![]);
-        }
-
-        let endpoint = self
-            .decode_window(state, triplet)
-            .map_err(|e| GigasttError::Inference { source: e.into() })?;
-        state.pending_samples = 0;
-        let ts = now_timestamp();
-
-        // True utterance endpoints only — never the encoder window cap.
-        // Cap used to emit `final` as a "backstop", which voice assistants
-        // (Irene) treated as "command complete" mid-phrase; it now commits a
-        // stable prefix and emits a non-final partial instead.
-        let speech_endpoint = Self::speech_endpoint(
-            state.endpoint_mode,
-            endpoint,
-            vad_endpoint,
-            state.vad_endpointer.is_some(),
-        );
-
-        if speech_endpoint {
-            let reason = if vad_endpoint {
-                EndpointReason::Vad
-            } else {
-                EndpointReason::Blank
-            };
-            let mut seg = state.assembler.finalize_with_reason(ts, reason);
-            self.enrich_final_segment(&mut seg, state);
-            Self::slide_streaming_window(state);
-            if seg.text.trim().is_empty() {
-                return Ok(vec![]);
-            }
-            return Ok(vec![seg]);
-        }
-
-        if over_cap {
-            // Encoder cost bound: commit live words so they are not lost when
-            // the window slides, but do **not** end the utterance.
-            state.assembler.commit_live();
-            Self::slide_streaming_window(state);
-            if state.assembler.is_empty() {
-                return Ok(vec![]);
-            }
-            return Ok(vec![state.assembler.partial(ts)]);
-        }
-
-        if state.assembler.is_empty() {
-            return Ok(vec![]);
-        }
-        Ok(vec![state.assembler.partial(ts)])
-    }
-
-    /// Whether this decode should close the utterance (`final` / `speech_final`).
-    ///
-    /// Pure helper so endpoint policy is unit-testable without ONNX.
-    pub(crate) fn speech_endpoint(
-        mode: EndpointMode,
-        decoder_blank_endpoint: bool,
-        vad_endpoint: bool,
-        vad_attached: bool,
-    ) -> bool {
-        let decoder_endpoint = decoder_blank_endpoint && !vad_attached;
-        match mode {
-            EndpointMode::Auto => decoder_endpoint || vad_endpoint,
-            // Assistant: only VAD silence (when attached). Blank-run alone is too
-            // aggressive for multi-word voice commands.
-            EndpointMode::Assistant => vad_endpoint,
-            EndpointMode::Manual => false,
-        }
-    }
-
-    /// Slide the streaming audio window, retaining left context for the next decode.
-    fn slide_streaming_window(state: &mut StreamingState) {
-        let keep = STREAM_LEFT_CONTEXT_SAMPLES.min(state.audio_buffer.len());
-        let slide_off = state.audio_buffer.len() - keep;
-        if slide_off > 0 {
-            audio::consume_audio_buffer(&mut state.audio_buffer, slide_off);
-            state.window_start_samples += slide_off;
-        }
-        state.context_samples = keep;
-    }
-
-    /// Re-decode the whole retained window from a fresh decoder state and update
-    /// the assembler with the context-suppressed tail. Returns whether the
-    /// decoder detected an endpoint. Shared by [`Engine::process_chunk`]
-    /// (strided) and [`Engine::finish_stream`] (forced at end of stream).
-    fn decode_window(
-        &self,
-        state: &mut StreamingState,
-        triplet: &mut SessionTriplet,
-    ) -> anyhow::Result<bool> {
-        let mel_start = std::time::Instant::now();
-        let num_frames = self.features.compute_mel(
-            &state.audio_buffer,
-            &mut state.mel_fft_input,
-            &mut state.mel_power,
-            &mut state.mel_output,
-        );
-        tracing::debug!(
-            elapsed_us = mel_start.elapsed().as_micros() as u64,
-            "mel_compute"
-        );
-        if num_frames == 0 {
-            return Ok(false);
-        }
-
-        // Encoder-frame offset of the window start (drift-free: a single division
-        // over the cumulative slid-off sample count).
-        let frame_offset = state.window_start_samples / (HOP_LENGTH * ENCODER_SUBSAMPLING);
-
-        // The window overlaps the previous one, so persisting the LSTM state
-        // would double-condition the prediction network — decode fresh.
-        let mut decoder_state = DecoderState::new(self.tokenizer.blank_id());
-        // Streaming always uses the engine boot biaser (per-request hotwords
-        // apply to file-transcription paths that carry TranscribeOverrides).
-        let (all_words, endpoint) = self.run_inference(
-            triplet,
-            &state.mel_output[..],
-            num_frames,
-            &mut decoder_state,
-            frame_offset,
-            true, // streaming: ANE low-latency pad floor when available
-            self.biaser.as_ref(),
-        )?;
-
-        // Suppress words inside the already-emitted left context so a slid
-        // window does not re-emit committed words.
-        let window_start_s = frame_offset as f64 * SECONDS_PER_FRAME;
-        let context_boundary_s = window_start_s + state.context_samples as f64 / 16000.0;
-        #[cfg_attr(not(feature = "diarization"), allow(unused_mut))]
-        let mut tail: Vec<WordInfo> = all_words
-            .into_iter()
-            .filter(|w| w.start + f64::EPSILON >= context_boundary_s)
-            .collect();
-
-        #[cfg(feature = "diarization")]
-        if let Some(dia) = state.diarization_state.as_mut()
-            && let Some(speaker) = diarization::last_turn_speaker(dia)
-        {
-            for w in &mut tail {
-                w.speaker = Some(speaker);
-            }
-        }
-
-        state.assembler.set_words(tail);
-        Ok(endpoint)
-    }
-
-    /// Decode any audio buffered since the last strided decode, then finalize.
-    /// Call when the stream ends (Stop / EOF) so the decode-stride batching does
-    /// not drop trailing words. Best-effort: on decode failure, falls back to a
-    /// plain flush of whatever the assembler already holds.
-    pub fn finish_stream(
-        &self,
-        state: &mut StreamingState,
-        triplet: &mut SessionTriplet,
-    ) -> Option<TranscriptSegment> {
-        let has_pending = state.pending_samples > 0 && state.audio_buffer.len() >= N_FFT;
-        if has_pending && let Err(e) = self.decode_window(state, triplet) {
-            tracing::warn!("finish_stream decode failed: {e:#}");
-        }
-        self.flush_state(state)
-    }
-
-    /// Flush accumulated text as a Final segment (called on Stop/Close).
-    pub fn flush_state(&self, state: &mut StreamingState) -> Option<TranscriptSegment> {
-        if state.assembler.is_empty() {
-            return None;
-        }
-        let mut seg = state
-            .assembler
-            .finalize_with_reason(now_timestamp(), EndpointReason::Stop);
-        self.enrich_final_segment(&mut seg, state);
-        Some(seg)
-    }
-
-    /// Post-process a finalized streaming segment: ITN, then punctuation/casing
-    /// restoration on the joined `text`. Mirrors
-    /// [`Engine::finish_transcribe_result`]'s policy exactly — the per-session
-    /// override wins over the engine boot default (`None` keeps it), and the
-    /// `punctuator` guard makes the pass a graceful no-op when no punct model
-    /// is attached. Word payloads keep the raw decoder output, exactly like the
-    /// file path. Runs only at finalization boundaries (endpoint flush /
-    /// Stop-flush), so `partial` payloads are never rewritten and live previews
-    /// don't flicker between hypotheses.
-    ///
-    /// Latency: measured via `punctuation::tests::test_restore_latency_short_segments`
-    /// (debug build, Apple Silicon), `restore` costs p95 ≈ 0.45–1.0 ms on 1–10
-    /// word segments — roughly two orders of magnitude below the 100 ms budget
-    /// that would force a segment-length gate, so enrichment always runs
-    /// regardless of segment length.
-    fn enrich_final_segment(&self, seg: &mut TranscriptSegment, state: &StreamingState) {
-        let text = std::mem::take(&mut seg.text);
-        seg.text = self.apply_text_postprocess(
-            text,
-            state.itn.unwrap_or(self.itn),
-            state.punctuation.unwrap_or(true),
-        );
-    }
-
-    /// Transcribe an audio file to text (supports WAV, MP3, M4A/AAC, OGG, FLAC).
-    ///
-    /// Decodes the file to mono 16kHz, runs the full encoder+decoder pipeline,
-    /// and returns the recognized text with word-level details and duration.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`GigasttError::InvalidAudio`] if the file cannot be decoded, or
-    /// [`GigasttError::Inference`] if the ONNX runtime fails.
-    /// Unified file-transcription entry point. Prefer this over the
-    /// combinatorial `transcribe_*` wrappers when building new call sites.
-    ///
-    /// Routes on [`TranscribeRequest::source`]:
-    /// - [`TranscribeSource::Path`] / [`TranscribeSource::Bytes`] — decode then
-    ///   mono pipeline (optional diarization)
-    /// - [`TranscribeSource::Samples`] — mono pipeline on pre-decoded audio
-    /// - [`TranscribeSource::Channels`] — per-channel decode + merge (diarization
-    ///   flag ignored; channel index is the speaker label)
-    pub fn transcribe_request(
-        &self,
-        req: TranscribeRequest<'_>,
-        triplet: &mut SessionTriplet,
-    ) -> Result<TranscribeResult, GigasttError> {
-        use std::sync::atomic::Ordering::Relaxed;
-        // Bridge the request's shared cancellation / progress handles into the
-        // lightweight closures the decode chain consults. Both own a clone of
-        // the `Arc`, so they outlive the borrow of `req` and stay valid for the
-        // whole call. Absent handles leave `ctl` all-`None`, and every decode
-        // function then runs its historical, byte-identical path.
-        let abort_fn: Option<Box<dyn Fn() -> bool>> = req.abort.as_ref().map(|flag| {
-            let flag = flag.clone();
-            Box::new(move || flag.load(Relaxed)) as Box<dyn Fn() -> bool>
-        });
-        let progress_fn: Option<Box<dyn Fn(u64)>> = req.progress.as_ref().map(|counter| {
-            let counter = counter.clone();
-            Box::new(move |n: u64| counter.store(n, Relaxed)) as Box<dyn Fn(u64)>
-        });
-        let ctl = DecodeControls {
-            abort: abort_fn.as_deref(),
-            on_progress: progress_fn.as_deref(),
-        };
-
-        // Opt-in operator length limit (`--max-audio-secs`); `None` = unlimited.
-        // The streaming path honors it verbatim; the whole-buffer decoders clamp
-        // it down to their fixed safety ceiling.
-        let max_audio_secs = req.max_audio_secs;
-
-        match req.source {
-            #[cfg(feature = "file-decode")]
-            TranscribeSource::Path(path) => {
-                if self.stream_eligible(req.diarization) {
-                    self.transcribe_stream_file(
-                        |spec| {
-                            audio::FileWindows::open(path, spec, max_audio_secs)
-                                .map_err(audio::decode_error)
-                        },
-                        triplet,
-                        &req.overrides,
-                        req.hotwords,
-                        ctl,
-                    )
-                } else {
-                    let float_samples = audio::decode_audio_file_bounded(path, max_audio_secs)
-                        .map_err(audio::decode_error)?;
-                    self.transcribe_samples_with_overrides(
-                        &float_samples,
-                        triplet,
-                        &req.overrides,
-                        req.hotwords,
-                        req.diarization,
-                        req.diarization_outcome.as_deref(),
-                        ctl,
-                    )
-                }
-            }
-            #[cfg(feature = "file-decode")]
-            TranscribeSource::Bytes(data) => {
-                if self.stream_eligible(req.diarization) {
-                    self.transcribe_stream_file(
-                        |spec| {
-                            audio::FileWindows::from_bytes(data.clone(), spec, max_audio_secs)
-                                .map_err(audio::decode_error)
-                        },
-                        triplet,
-                        &req.overrides,
-                        req.hotwords,
-                        ctl,
-                    )
-                } else {
-                    let float_samples =
-                        audio::decode_audio_bytes_shared_bounded(data, max_audio_secs)
-                            .map_err(audio::decode_error)?;
-                    self.transcribe_samples_with_overrides(
-                        &float_samples,
-                        triplet,
-                        &req.overrides,
-                        req.hotwords,
-                        req.diarization,
-                        req.diarization_outcome.as_deref(),
-                        ctl,
-                    )
-                }
-            }
-            TranscribeSource::Samples(samples) => self.transcribe_samples_with_overrides(
-                samples,
-                triplet,
-                &req.overrides,
-                req.hotwords,
-                req.diarization,
-                req.diarization_outcome.as_deref(),
-                ctl,
-            ),
-            #[cfg(feature = "file-decode")]
-            TranscribeSource::ChannelStreams { data, channels } => self.transcribe_channel_streams(
-                data,
-                channels,
-                max_audio_secs,
-                triplet,
-                &req.overrides,
-                req.hotwords,
-                ctl.abort_only(),
-            ),
-            TranscribeSource::Channels(channels) => self.transcribe_channels_inner(
-                channels,
-                triplet,
-                &req.overrides,
-                req.hotwords,
-                ctl.abort_only(),
-            ),
-        }
-    }
-
-    /// True when a `Path` / `Bytes` request can be served by the windowed
-    /// streaming decode, whose peak audio memory is O(one window) rather than
-    /// O(file).
-    ///
-    /// Diarization embeds the whole clip in a second pass, so it cannot run
-    /// against a stream that is never fully resident: it still forces the
-    /// whole-buffer decode, and with it the duration ceiling. VAD does not — it
-    /// is causal, and [`VadWindows`](audio::VadWindows) runs it inside the
-    /// stream. `diarize` only matters with the `diarization` feature compiled
-    /// in.
-    #[cfg(feature = "file-decode")]
-    fn stream_eligible(&self, diarize: bool) -> bool {
-        !(cfg!(feature = "diarization") && diarize)
-    }
-
-    /// Streaming mono file transcription, with the VAD stage in front when one
-    /// is attached and the request did not opt out.
-    ///
-    /// `open` builds a fresh window source for a given geometry. It is called
-    /// once, or twice when the VAD stage declines — the model failed mid-stream,
-    /// or the scan found no speech at all in a non-empty clip — in which case
-    /// the clip is decoded whole, exactly as the batch path did.
-    #[cfg(feature = "file-decode")]
-    fn transcribe_stream_file(
-        &self,
-        open: impl Fn(WindowSpec) -> Result<audio::FileWindows, GigasttError>,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-        ctl: DecodeControls,
-    ) -> Result<TranscribeResult, GigasttError> {
-        let use_vad = self.vad.is_some() && overrides.vad.unwrap_or(true);
-        if let (true, Some(vad)) = (use_vad, &self.vad) {
-            let wall_start = std::time::Instant::now();
-            let request_biaser = hotwords.and_then(|hw| self.build_request_biaser(hw));
-            let biaser = self.select_biaser(hotwords, &request_biaser);
-            let mut windows = audio::VadWindows::new(
-                open(audio::VadWindows::pull_spec())?,
-                vad,
-                &self.vad_config,
-                window_spec(self.ane_encoder, self.variant.is_ctc()),
-                ctl.abort,
-            );
-            let mut words = self.decode_words_streaming(&mut windows, triplet, biaser, ctl)?;
-            if !windows.needs_fallback() {
-                // Words are decoded on the compressed (silence-removed)
-                // timeline; put them back on the clip's own.
-                let regions = windows.regions();
-                for w in &mut words {
-                    w.start = crate::vad::remap_compressed_seconds(w.start, regions, 16000.0);
-                    w.end = crate::vad::remap_compressed_seconds(w.end, regions, 16000.0);
-                }
-                let duration_s = windows.total_16k_samples() as f64 / 16000.0;
-                let wall_s = wall_start.elapsed().as_secs_f64();
-                tracing::info!(
-                    audio_s = format_args!("{duration_s:.2}"),
-                    wall_s = format_args!("{wall_s:.2}"),
-                    rtf = format_args!(
-                        "{:.3}",
-                        if duration_s > 0.0 {
-                            wall_s / duration_s
-                        } else {
-                            0.0
-                        }
-                    ),
-                    regions = regions.len(),
-                    "transcribe complete (streaming windows, vad)"
-                );
-                return Ok(self.finish_transcribe_result(words, duration_s, overrides));
-            }
-            // Either the VAD found no speech at all — tone or continuous speech
-            // against a bad threshold — or the model failed mid-stream (already
-            // logged with the cause). Both re-read the clip and decode it whole
-            // rather than returning an empty transcript.
-            tracing::warn!("VAD produced no usable speech regions; decoding full audio");
-        }
-        self.transcribe_stream_mono(
-            open(window_spec(self.ane_encoder, self.variant.is_ctc()))?,
-            triplet,
-            overrides,
-            hotwords,
-            ctl,
-        )
-    }
-
-    /// Mono file-transcription tail that pulls windows straight from the
-    /// container instead of decoding the whole file first.
-    ///
-    /// Equivalent to [`Engine::transcribe_samples_with_overrides`] on the
-    /// no-VAD, no-diarization path: [`FileWindows`](audio::FileWindows) yields
-    /// exactly the geometry [`SliceWindows`] would over the fully-decoded buffer
-    /// (one window for a single-pass-length stream, overlapping windows beyond),
-    /// so [`Engine::decode_words_streaming`] produces the same words — but peak
-    /// audio memory no longer scales with duration. The duration cap and its
-    /// exact error string are enforced inside the decode as before.
-    #[cfg(feature = "file-decode")]
-    fn transcribe_stream_mono(
-        &self,
-        mut windows: audio::FileWindows,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-        ctl: DecodeControls,
-    ) -> Result<TranscribeResult, GigasttError> {
-        let wall_start = std::time::Instant::now();
-
-        // Hotword biaser: engine boot biaser, temporary per-request, or off.
-        let request_biaser = hotwords.and_then(|hw| self.build_request_biaser(hw));
-        let biaser = self.select_biaser(hotwords, &request_biaser);
-
-        let words = self.decode_words_streaming(&mut windows, triplet, biaser, ctl)?;
-        // Exact once every window is consumed (the loop above drains to EOF).
-        let duration_s = windows.total_16k_samples() as f64 / 16000.0;
-        let result = self.finish_transcribe_result(words, duration_s, overrides);
-
-        let wall_s = wall_start.elapsed().as_secs_f64();
-        let rtf = if duration_s > 0.0 {
-            wall_s / duration_s
-        } else {
-            0.0
-        };
-        tracing::info!(
-            audio_s = format_args!("{duration_s:.2}"),
-            wall_s = format_args!("{wall_s:.2}"),
-            rtf = format_args!("{rtf:.3}"),
-            "transcribe complete (streaming windows)"
-        );
-
-        Ok(result)
-    }
-
-    /// Transcribe an audio file to text (supports WAV, MP3, M4A/AAC, OGG, FLAC).
-    ///
-    /// Decodes the file to mono 16kHz, runs the full encoder+decoder pipeline,
-    /// and returns the recognized text with word-level details and duration.
-    ///
-    /// Thin wrapper over [`Engine::transcribe_request`].
-    ///
-    /// # Errors
-    ///
-    /// Returns [`GigasttError::InvalidAudio`] if the file cannot be decoded, or
-    /// [`GigasttError::Inference`] if the ONNX runtime fails.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_file(
-        &self,
-        path: &str,
-        triplet: &mut SessionTriplet,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_request(
-            TranscribeRequest::new(TranscribeSource::Path(path)),
-            triplet,
-        )
-    }
-
-    /// Like [`Engine::transcribe_file`] but applies per-request recognition-knob
-    /// [`overrides`](TranscribeOverrides). With `TranscribeOverrides::default()`
-    /// this is byte-for-byte [`Engine::transcribe_file`]; the plain method
-    /// delegates here so binding call sites (FFI / UniFFI / Node) keep the
-    /// no-override signature unchanged.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_file_with_overrides(
-        &self,
-        path: &str,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_file_with_overrides_hotwords(path, triplet, overrides, None)
-    }
-
-    /// Like [`Engine::transcribe_file_with_overrides`] with optional per-request
-    /// [`HotwordOverride`] (semver-additive sibling so the no-hotwords signature
-    /// stays byte-stable).
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_file_with_overrides_hotwords(
-        &self,
-        path: &str,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_request(
-            TranscribeRequest::new(TranscribeSource::Path(path))
-                .with_overrides(*overrides)
-                .with_hotwords(hotwords),
-            triplet,
-        )
-    }
-
-    /// Transcribe audio from raw bytes in memory (no temp file needed).
-    ///
-    /// Backwards-compatible shim: clones `data` into a [`bytes::Bytes`] and
-    /// delegates to [`Engine::transcribe_bytes_shared`]. Prefer the shared
-    /// variant on hot paths (REST/SSE) to avoid the extra copy.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_bytes(
-        &self,
-        data: &[u8],
-        triplet: &mut SessionTriplet,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_bytes_shared(bytes::Bytes::copy_from_slice(data), triplet)
-    }
-
-    /// Transcribe audio from a reference-counted [`bytes::Bytes`] buffer
-    /// without cloning.
-    ///
-    /// Reuses the same decode/inference pipeline as [`Engine::transcribe_bytes`]
-    /// but hands the buffer straight to symphonia via [`audio::decode_audio_bytes_shared`].
-    /// This is the zero-copy entry point used by the REST upload handler so a
-    /// 50 MiB `axum::body::Bytes` body stays as a single in-memory buffer
-    /// instead of being cloned into a `Vec<u8>` before decode.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_bytes_shared(
-        &self,
-        data: bytes::Bytes,
-        triplet: &mut SessionTriplet,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_bytes_shared_with_overrides(data, triplet, &TranscribeOverrides::default())
-    }
-
-    /// Like [`Engine::transcribe_bytes_shared`] but applies per-request
-    /// recognition-knob [`overrides`](TranscribeOverrides). With
-    /// `TranscribeOverrides::default()` this is byte-for-byte
-    /// [`Engine::transcribe_bytes_shared`]; the plain method delegates here so
-    /// the zero-copy REST call site can opt into overrides without changing the
-    /// no-override signature that other callers rely on.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_bytes_shared_with_overrides(
-        &self,
-        data: bytes::Bytes,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_bytes_shared_with_overrides_hotwords(data, triplet, overrides, None)
-    }
-
-    /// Like [`Engine::transcribe_bytes_shared_with_overrides`] with optional
-    /// per-request [`HotwordOverride`].
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_bytes_shared_with_overrides_hotwords(
-        &self,
-        data: bytes::Bytes,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_request(
-            TranscribeRequest::new(TranscribeSource::Bytes(data))
-                .with_overrides(*overrides)
-                .with_hotwords(hotwords),
-            triplet,
-        )
-    }
-
-    /// Like [`Engine::transcribe_bytes_shared_with_overrides`], but also runs
-    /// offline speaker diarization (labels each word's `speaker`) when a speaker
-    /// encoder is loaded. Diarization is **opt-in per request** (`?diarization=true`
-    /// on the REST surface): the non-diarized transcribe methods never label
-    /// speakers, so a plain transcript — and the `channels=split` dual-mono
-    /// fallback — carries no speaker labels. Without the `diarization` feature or a
-    /// loaded speaker encoder this is byte-for-byte the non-diarized method.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_bytes_shared_with_overrides_diarized(
-        &self,
-        data: bytes::Bytes,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_bytes_shared_with_overrides_diarized_hotwords(
-            data, triplet, overrides, None,
-        )
-    }
-
-    /// Like [`Engine::transcribe_bytes_shared_with_overrides_diarized`] with
-    /// optional per-request [`HotwordOverride`].
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_bytes_shared_with_overrides_diarized_hotwords(
-        &self,
-        data: bytes::Bytes,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_request(
-            TranscribeRequest::new(TranscribeSource::Bytes(data))
-                .with_overrides(*overrides)
-                .with_hotwords(hotwords)
-                .with_diarization(true),
-            triplet,
-        )
-    }
-
-    /// Transcribe a multi-channel recording with one speaker label per channel.
-    ///
-    /// Runs the engine once per channel sequentially on the supplied triplet. The
-    /// caller is responsible for deciding whether to use this mode (e.g. after
-    /// checking for dual-mono). Channel 0 becomes `speaker_0`, channel 1
-    /// `speaker_1`, and so on. Results are merged into a single chronologically
-    /// ordered transcript.
-    ///
-    /// Per-channel `text` fields are ignored: the merged transcript's text is
-    /// rebuilt from the merged words after the final ITN/punctuation pass.
-    ///
-    /// Thin wrapper over [`Engine::transcribe_request`] with default overrides
-    /// and no hotwords.
-    #[cfg(feature = "file-decode")]
-    pub fn transcribe_channels(
-        &self,
-        channels: &[Vec<f32>],
-        triplet: &mut SessionTriplet,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_request(
-            TranscribeRequest::new(TranscribeSource::Channels(channels)),
-            triplet,
-        )
-    }
-
-    /// Per-channel decode + merge used by [`TranscribeSource::Channels`].
-    fn transcribe_channels_inner(
-        &self,
-        channels: &[Vec<f32>],
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-        ctl: DecodeControls,
-    ) -> Result<TranscribeResult, GigasttError> {
-        if channels.is_empty() {
-            return Ok(TranscribeResult {
-                text: String::new(),
-                words: Vec::new(),
-                duration_s: 0.0,
-                confidence: None,
-            });
-        }
-
-        let mut per_channel = Vec::with_capacity(channels.len());
-        for channel_samples in channels {
-            let words =
-                self.decode_words_for_samples(channel_samples, triplet, overrides, hotwords, ctl)?;
-            let duration_s = channel_samples.len() as f64 / 16000.0;
-            per_channel.push(TranscribeResult {
-                confidence: aggregate_confidence(&words),
-                text: String::new(),
-                words,
-                duration_s,
-            });
-        }
-
-        let merged = merge_channel_results(per_channel);
-        Ok(self.finish_transcribe_result(merged.words, merged.duration_s, overrides))
-    }
-
-    /// Streaming twin of the `channels=split` decode.
-    ///
-    /// Each channel of `data` runs through the windowed loop the mono file path
-    /// uses, so
-    /// peak audio memory is one window instead of every channel of the whole
-    /// file — which is what kept channel splitting on a duration ceiling. The
-    /// per-channel results are merged exactly as the whole-buffer twin merges
-    /// them.
-    ///
-    /// The caller decides *whether* to split; use
-    /// [`scan_channels`](crate::inference::audio::scan_channels), which answers
-    /// that in one pass without materializing anything.
-    ///
-    /// No progress sink is threaded through: each channel restarts the sample
-    /// clock, so a shared monotonic counter would go backwards — the same
-    /// reason the whole-buffer twin passes `abort_only`.
-    // Bundling these behind `&TranscribeRequest` is what one would want, but the
-    // caller's match moves `data` out of `req.source`, so the request cannot be
-    // borrowed whole afterwards. Same shape as `run_inference` above.
-    #[allow(clippy::too_many_arguments)]
-    #[cfg(feature = "file-decode")]
-    fn transcribe_channel_streams(
-        &self,
-        data: bytes::Bytes,
-        channels: usize,
-        max_audio_secs: Option<f64>,
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-        ctl: DecodeControls,
-    ) -> Result<TranscribeResult, GigasttError> {
-        let request_biaser = hotwords.and_then(|hw| self.build_request_biaser(hw));
-        let biaser = self.select_biaser(hotwords, &request_biaser);
-
-        let spec = window_spec(self.ane_encoder, self.variant.is_ctc());
-        let mut per_channel = Vec::with_capacity(channels);
-        for k in 0..channels {
-            let mut windows =
-                audio::FileWindows::from_bytes_channel(data.clone(), spec, max_audio_secs, k)
-                    .map_err(audio::decode_error)?;
-            let words = self.decode_words_streaming(&mut windows, triplet, biaser, ctl)?;
-            per_channel.push(TranscribeResult {
-                confidence: aggregate_confidence(&words),
-                text: String::new(),
-                words,
-                duration_s: windows.total_16k_samples() as f64 / 16000.0,
-            });
-        }
-
-        let merged = merge_channel_results(per_channel);
-        Ok(self.finish_transcribe_result(merged.words, merged.duration_s, overrides))
-    }
-
-    /// Run the full mel + encoder + RNN-T decode pipeline on an already-decoded
-    /// 16 kHz f32 sample buffer. Shared tail of [`Engine::transcribe_request`]
-    /// for mono sources (and unit tests).
-    fn transcribe_samples(
-        &self,
-        float_samples: &[f32],
-        triplet: &mut SessionTriplet,
-    ) -> Result<TranscribeResult, GigasttError> {
-        self.transcribe_request(
-            TranscribeRequest::new(TranscribeSource::Samples(float_samples)),
-            triplet,
-        )
-    }
-
-    /// Override-aware tail of the file-transcription pipeline. With
-    /// `TranscribeOverrides::default()` (all `None`) it is byte-for-byte the
-    /// engine-default path; each `Some(_)` field flips the corresponding
-    /// post-processing knob for this call only. `overrides` is assumed already
-    /// validated by [`Engine::validate_overrides`] — an on-request with the
-    /// resource missing degrades gracefully (VAD absent → whole-buffer decode)
-    /// rather than erroring here.
-    /// Override-aware tail of the file-transcription pipeline. With
-    /// `TranscribeOverrides::default()` (all `None`) it is byte-for-byte the
-    /// engine-default path; each `Some(_)` field flips the corresponding
-    /// post-processing knob for this call only. `overrides` is assumed already
-    /// validated by [`Engine::validate_overrides`] — an on-request with the
-    /// resource missing degrades gracefully (VAD absent → whole-buffer decode)
-    /// rather than erroring here.
-    #[allow(clippy::too_many_arguments)] // overrides + diarization sink + controls; bundle later if it grows again
-    fn transcribe_samples_with_overrides(
-        &self,
-        float_samples: &[f32],
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-        diarize: bool,
-        diar_sink: Option<&std::sync::OnceLock<DiarizationOutcome>>,
-        ctl: DecodeControls,
-    ) -> Result<TranscribeResult, GigasttError> {
-        // `diarize` is opt-in per request: offline speaker diarization only runs
-        // when the caller asked for it (REST `?diarization=true`). A plain
-        // transcript — and the `channels=split` dual-mono fallback — must carry no
-        // speaker labels, so the default paths pass `false`.
-        let wall_start = std::time::Instant::now();
-        let duration_s = float_samples.len() as f64 / 16000.0;
-
-        #[cfg_attr(not(feature = "diarization"), allow(unused_mut))]
-        let mut words =
-            self.decode_words_for_samples(float_samples, triplet, overrides, hotwords, ctl)?;
-
-        // Record *why* speakers were or were not labeled into the caller's sink
-        // so a `?diarization=true` request that produced no labels can be
-        // surfaced with a reason instead of an all-empty-speaker transcript.
-        #[cfg(feature = "diarization")]
-        if diarize {
-            let outcome = match self
-                .speaker_encoder
-                .as_ref()
-                .and_then(|lazy| lazy.get_or_load())
-            {
-                None => DiarizationOutcome::NoSpeakerModel,
-                Some(enc) => match diarization::run_offline(&enc, float_samples) {
-                    Ok(turns) => {
-                        diarization::assign_speakers_by_midpoint(&turns, &mut words);
-                        DiarizationOutcome::Applied
-                    }
-                    Err(declined) => declined,
-                },
-            };
-            if let Some(sink) = diar_sink {
-                let _ = sink.set(outcome);
-            }
-        }
-        // A build compiled without the `diarization` feature can never label
-        // speakers; report that per request rather than silently dropping the flag.
-        #[cfg(not(feature = "diarization"))]
-        if diarize && let Some(sink) = diar_sink {
-            let _ = sink.set(DiarizationOutcome::NoSpeakerModel);
-        }
-
-        let result = self.finish_transcribe_result(words, duration_s, overrides);
-
-        let wall_s = wall_start.elapsed().as_secs_f64();
-        let rtf = if duration_s > 0.0 {
-            wall_s / duration_s
-        } else {
-            0.0
-        };
-        let encoder_label = if self.int8 { "int8" } else { "fp32" };
-        let backend_label = if cfg!(feature = "candle") {
-            "candle"
-        } else if cfg!(feature = "ane") {
-            "ane"
-        } else if cfg!(feature = "coreml") {
-            "coreml"
-        } else if cfg!(feature = "cuda") {
-            "cuda"
-        } else {
-            "cpu"
-        };
-        tracing::info!(
-            audio_s = format_args!("{duration_s:.2}"),
-            wall_s = format_args!("{wall_s:.2}"),
-            rtf = format_args!("{rtf:.3}"),
-            encoder = format_args!("{encoder_label}/{backend_label}"),
-            "transcribe complete"
-        );
-
-        Ok(result)
-    }
-
-    /// Decode a 16 kHz f32 buffer to words: single-pass for short inputs (one
-    /// encoder Run over the whole buffer), chunked overlapping windows for long
-    /// inputs so encoder activation memory stays O(chunk), not O(file). Both
-    /// paths produce the same `Vec<WordInfo>` shape. This is the no-VAD path and
-    /// the per-region decode used by [`Engine::decode_speech_regions`].
-    ///
-    /// `biaser` is the effective hotword biaser for this call (engine boot
-    /// biaser, a temporary per-request biaser, or `None` when forced off).
-    fn decode_words(
-        &self,
-        samples: &[f32],
-        triplet: &mut SessionTriplet,
-        biaser: Option<&bias::Biaser>,
-        ctl: DecodeControls,
-    ) -> Result<Vec<WordInfo>, GigasttError> {
-        let spec = window_spec(self.ane_encoder, self.variant.is_ctc());
-        if spec.is_single_pass(samples.len()) {
-            // A single-pass decode is one encoder Run with no window loop to
-            // check between, so honour cancellation before starting it. The
-            // caller's watchdog treats this whole clip as one "window".
-            if ctl.aborted() {
-                return Err(GigasttError::Cancelled);
-            }
-            let (features, num_frames) = self.features.compute(samples);
-            tracing::info!("Extracted {} mel frames", num_frames);
-            let mut decoder_state = DecoderState::new(self.tokenizer.blank_id());
-            let words = self
-                .run_inference(
-                    triplet,
-                    &features,
-                    num_frames,
-                    &mut decoder_state,
-                    0,
-                    false, // file-mode fill floor
-                    biaser,
-                )
-                .map_err(|e| GigasttError::Inference { source: e.into() })?
-                .0;
-            ctl.report(samples.len() as u64);
-            Ok(words)
-        } else {
-            tracing::info!(
-                "Long-form chunked decode: {:.1}s in ~{}s windows ({}s overlap, ane={})",
-                samples.len() as f64 / 16000.0,
-                spec.window() / 16000,
-                spec.overlap() / 16000,
-                self.ane_encoder,
-            );
-            self.decode_words_streaming(&mut SliceWindows::new(samples, spec), triplet, biaser, ctl)
-        }
-    }
-
-    /// Decode only the VAD-detected speech `regions` of `float_samples`: copy the
-    /// speech spans into one silence-free buffer, decode it, then remap each
-    /// word's start/end from the compressed (silence-removed) timeline back to
-    /// the original timeline via [`crate::vad::remap_compressed_seconds`]. Empty
-    /// `regions` (no speech) yields no words.
-    fn decode_speech_regions(
-        &self,
-        float_samples: &[f32],
-        regions: &[(usize, usize)],
-        triplet: &mut SessionTriplet,
-        biaser: Option<&bias::Biaser>,
-        ctl: DecodeControls,
-    ) -> Result<Vec<WordInfo>, GigasttError> {
-        if regions.is_empty() {
-            tracing::info!("VAD found no speech; skipping decode");
-            return Ok(Vec::new());
-        }
-        let speech_len: usize = regions.iter().map(|(s, e)| e - s).sum();
-        let mut speech = Vec::with_capacity(speech_len);
-        for &(s, e) in regions {
-            speech.extend_from_slice(&float_samples[s..e]);
-        }
-        tracing::info!(
-            "VAD kept {}/{} samples ({} speech region(s))",
-            speech_len,
-            float_samples.len(),
-            regions.len()
-        );
-        let mut words = self.decode_words(&speech, triplet, biaser, ctl)?;
-        for w in &mut words {
-            w.start = crate::vad::remap_compressed_seconds(w.start, regions, 16000.0);
-            w.end = crate::vad::remap_compressed_seconds(w.end, regions, 16000.0);
-        }
-        Ok(words)
-    }
-
-    /// Long-form decode: pull overlapping windows from `windows`, encode and
-    /// decode each independently with a fresh [`DecoderState`], offset each
-    /// chunk's word timestamps by the chunk's absolute start, then stitch the
-    /// per-chunk word lists with overlap de-dup via [`stitch_chunk_words`].
-    ///
-    /// Peak encoder activation memory is bounded by the source's window length
-    /// (24s ort / 30s ANE, see [`super::windows::chunk_window_samples`]) rather than the full
-    /// file length. Window starts are aligned to encoder frame boundaries
-    /// (multiples of `HOP_LENGTH * ENCODER_SUBSAMPLING`) so the per-chunk frame
-    /// offset is exact, matching the streaming path's math.
-    ///
-    /// Takes the PCM as a [`PcmWindows`] source rather than a `&[f32]`, so the
-    /// loop makes no assumption that the whole file is in memory.
-    fn decode_words_streaming(
-        &self,
-        windows: &mut dyn PcmWindows,
-        triplet: &mut SessionTriplet,
-        biaser: Option<&bias::Biaser>,
-        ctl: DecodeControls,
-    ) -> Result<Vec<WordInfo>, GigasttError> {
-        let overlap = windows.spec().overlap();
-        let frame_samples = HOP_LENGTH * ENCODER_SUBSAMPLING;
-
-        let mut merged: Vec<WordInfo> = Vec::new();
-        while let Some(window) = windows.next_window()? {
-            // Cooperative cancellation checkpoint: a flipped abort flag ends the
-            // run at this window boundary, so a cancelled request (client
-            // disconnect, `DELETE /v1/jobs/{id}`, shutdown, or the no-progress
-            // watchdog) frees its pooled session within one window instead of
-            // decoding the rest of the file.
-            if ctl.aborted() {
-                return Err(GigasttError::Cancelled);
-            }
-            let start = window.start_sample;
-            // Window ends advance monotonically, so this doubles as the
-            // cumulative processed-sample count reported below.
-            let win_end = start + window.samples.len();
-            let (features, num_frames) = self.features.compute(window.samples);
-            let frame_offset = start / frame_samples;
-            let mut decoder_state = DecoderState::new(self.tokenizer.blank_id());
-            let (chunk_words, _endpoint) = self
-                .run_inference(
-                    triplet,
-                    &features,
-                    num_frames,
-                    &mut decoder_state,
-                    frame_offset,
-                    false, // file-mode fill floor
-                    biaser,
-                )
-                .map_err(|e| GigasttError::Inference { source: e.into() })?;
-
-            // Seam between the previous chunk's window and this one falls at the
-            // midpoint of their overlap region, in absolute seconds.
-            let overlap_mid_s = (start as f64 + overlap as f64 / 2.0) / 16000.0;
-            merged = stitch_chunk_words(merged, chunk_words, overlap_mid_s);
-
-            // A completed window is one unit of real progress: it resets the
-            // server's no-progress deadline and advances a job's bar by the
-            // seconds of audio just decoded (`win_end / 16000`).
-            ctl.report(win_end as u64);
-        }
-        Ok(merged)
-    }
-
-    /// Decode a 16 kHz f32 buffer to words, applying VAD if configured.
-    ///
-    /// The VAD path is taken only when a VAD is attached AND the caller hasn't
-    /// opted out via `overrides.vad`. `?vad=false` forces whole-buffer decode
-    /// even on a VAD-enabled engine; `None` uses the engine default (VAD path
-    /// iff a VAD is attached). `vad = Some(true)` on a VAD-less engine can't
-    /// reach here — callers should validate overrides first — but the
-    /// `self.vad.is_some()` guard keeps this correct regardless.
-    ///
-    /// Hotword selection via `hotwords`:
-    /// - `None` → engine boot biaser
-    /// - `Some(empty)` → force biasing off
-    /// - `Some(phrases)` → temporary biaser built for this request only
-    fn decode_words_for_samples(
-        &self,
-        float_samples: &[f32],
-        triplet: &mut SessionTriplet,
-        overrides: &TranscribeOverrides,
-        hotwords: Option<&HotwordOverride>,
-        ctl: DecodeControls,
-    ) -> Result<Vec<WordInfo>, GigasttError> {
-        // Temporary biaser only when the request supplies hotwords. Owned
-        // here so the `Option<&Biaser>` passed into decode stays valid for
-        // the whole call without cloning the engine's boot biaser.
-        let request_biaser = hotwords.and_then(|hw| self.build_request_biaser(hw));
-        let biaser = self.select_biaser(hotwords, &request_biaser);
-
-        let use_vad = self.vad.is_some() && overrides.vad.unwrap_or(true);
-        match (use_vad, &self.vad) {
-            (true, Some(vad)) => {
-                match vad.speech_regions_with_abort(float_samples, &self.vad_config, ctl.abort) {
-                    Ok(regions) if regions.is_empty() => {
-                        // Tone / continuous speech can yield zero regions on a bad
-                        // threshold; fall back to fixed-window / full decode rather
-                        // than returning an empty transcript (lab T-045).
-                        tracing::warn!(
-                            "VAD found no speech regions; falling back to full/chunked decode"
-                        );
-                        self.decode_words(float_samples, triplet, biaser, ctl)
-                    }
-                    Ok(regions) => {
-                        self.decode_speech_regions(float_samples, &regions, triplet, biaser, ctl)
-                    }
-                    Err(e) => {
-                        // A flipped abort flag stops the VAD scan too; surface it
-                        // as cancellation instead of silently re-decoding the whole
-                        // file, which would ignore the cancel and keep the triplet.
-                        if ctl.aborted() {
-                            return Err(GigasttError::Cancelled);
-                        }
-                        tracing::warn!("VAD failed, decoding full audio: {e:#}");
-                        self.decode_words(float_samples, triplet, biaser, ctl)
-                    }
-                }
-            }
-            _ => self.decode_words(float_samples, triplet, biaser, ctl),
-        }
-    }
-
-    /// Build the final [`TranscribeResult`] from raw words: join text, apply ITN,
-    /// and apply punctuation restoration. Word-level timing is left untouched.
-    /// Per-request overrides win over engine defaults; a `None` override keeps
-    /// the boot policy.
-    fn finish_transcribe_result(
-        &self,
-        words: Vec<WordInfo>,
-        duration_s: f64,
-        overrides: &TranscribeOverrides,
-    ) -> TranscribeResult {
-        let text: String = words
-            .iter()
-            .map(|w| w.word.as_str())
-            .collect::<Vec<_>>()
-            .join(" ");
-
-        // Optional ITN then punctuation. Per-request override wins over the
-        // engine default; `None` keeps the boot policy. Word-level timing is
-        // left untouched — only the joined `text` is rewritten.
-        let text = self.apply_text_postprocess(
-            text,
-            overrides.itn.unwrap_or(self.itn),
-            overrides.punctuation.unwrap_or(true),
-        );
-
-        TranscribeResult {
-            text,
-            confidence: aggregate_confidence(&words),
-            words,
-            duration_s,
-        }
-    }
-
-    /// Run encoder + decode. `low_latency` selects the streaming encoder path
-    /// ([`RuntimeSession::run_low_latency`]) so ANE can pad underfilled short
-    /// windows; file mode keeps the calibrated 0.5 fill floor. `biaser` is the
-    /// effective per-call hotword biaser (boot, request override, or off).
-    #[allow(clippy::too_many_arguments)] // encoder/decode call site; bundle later if it grows again
-    fn run_inference(
-        &self,
-        triplet: &mut SessionTriplet,
-        features: &[f32],
-        num_frames: usize,
-        decoder_state: &mut DecoderState,
-        frame_offset: usize,
-        low_latency: bool,
-        biaser: Option<&bias::Biaser>,
-    ) -> anyhow::Result<(Vec<WordInfo>, bool)> {
-        // Reuse the encoder input tensors: resize the signal tensor to the
-        // current frame count and overwrite both buffers in place.
-        triplet.encoder_inputs[0].resize_to(Shape::new(vec![1, N_MELS, num_frames]));
-        triplet.encoder_inputs[0]
-            .as_f32_mut()
-            .context("encoder signal tensor is not f32")?
-            .copy_from_slice(features);
-        triplet.encoder_inputs[1]
-            .as_i64_mut()
-            .context("encoder length tensor is not i64")?[0] = num_frames as i64;
-
-        let enc_start = std::time::Instant::now();
-        let encoder_outputs = if low_latency {
-            triplet
-                .encoder
-                .run_low_latency(&triplet.encoder_inputs)
-                .context("Encoder inference failed")?
-        } else {
-            triplet
-                .encoder
-                .run(&triplet.encoder_inputs)
-                .context("Encoder inference failed")?
-        };
-        tracing::info!(
-            elapsed_ms = enc_start.elapsed().as_millis() as u64,
-            "encoder_inference"
-        );
-
-        let enc_len = match encoder_outputs[1].view().data() {
-            TensorDataView::I32(v) => usize::try_from(v[0]).context("Negative encoder length")?,
-            TensorDataView::I64(v) => usize::try_from(v[0]).context("Negative encoder length")?,
-            _ => anyhow::bail!("Unexpected encoder length tensor type"),
-        };
-
-        tracing::debug!("Encoder output: {} frames", enc_len);
-
-        // CTC head: the single encoder emits per-frame class log-probs
-        // (`[1, T', 71]`, row-major). Decode them directly — there is no
-        // prediction network / joiner, so we return before the RNN-T block
-        // borrows `encoder_outputs` for the decode loop.
-        //
-        // A glossary switches the decode to a prefix beam, which is the only
-        // form that can act on one: a per-frame argmax has no continuation
-        // state to steer. Without hotwords the greedy path runs untouched, so
-        // output for everyone else is byte-for-byte what it was.
-        if self.variant.is_ctc() {
-            let log_probs = encoder_outputs[0]
-                .view()
-                .data()
-                .as_f32()
-                .context("CTC log_probs tensor is not f32")?;
-            let tokens = match biaser {
-                Some(b) => ctc::ctc_prefix_beam_decode(
-                    log_probs,
-                    enc_len,
-                    self.tokenizer.vocab_size(),
-                    self.tokenizer.blank_id(),
-                    b,
-                ),
-                None => ctc::ctc_greedy_decode(
-                    log_probs,
-                    enc_len,
-                    self.tokenizer.vocab_size(),
-                    self.tokenizer.blank_id(),
-                ),
-            };
-            let words = ctc::ctc_tokens_to_words(&self.tokenizer, &tokens, frame_offset);
-            return Ok((words, false)); // CTC has no endpoint signal
-        }
-
-        // RNN-T greedy decode — the encoder output is borrowed for the decode loop.
-        let dec_start = std::time::Instant::now();
-        let result = decode::greedy_decode(
-            triplet
-                .decoder
-                .as_deref()
-                .expect("RNN-T decoder session must be loaded for a non-CTC head"),
-            triplet
-                .joiner
-                .as_deref()
-                .expect("RNN-T joiner session must be loaded for a non-CTC head"),
-            &encoder_outputs[0].view(),
-            enc_len,
-            self.tokenizer.blank_id(),
-            decoder_state,
-            biaser,
-        )?;
-        tracing::info!(
-            elapsed_ms = dec_start.elapsed().as_millis() as u64,
-            "greedy_decode"
-        );
-
-        // Convert token infos to words with timestamps
-        let words = self.tokens_to_words(&result.tokens, frame_offset);
-
-        tracing::info!(
-            tokens = result.tokens.len(),
-            words = words.len(),
-            duration_ms = dec_start.elapsed().as_millis() as u64,
-            "Decoded tokens"
-        );
-
-        Ok((words, result.endpoint_detected))
-    }
-
-    /// Convert decoded tokens into words with timestamps and confidence.
-    fn tokens_to_words(&self, tokens: &[decode::TokenInfo], frame_offset: usize) -> Vec<WordInfo> {
-        TokenFormatter::tokens_to_words(&self.tokenizer, tokens, frame_offset)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    #[cfg(feature = "diarization")]
-    #[cfg(feature = "diarization")]
-    use crate::inference::diarization::SPEAKER_EMBEDDING_DIM;
-    use crate::inference::windows::CHUNK_THRESHOLD_SAMPLES;
-    use crate::inference::{EndpointMode, WordInfo};
-    #[cfg(feature = "diarization")]
-    use polyvoice::Embedder;
-
-    fn word(text: &str, start: f64, end: f64) -> WordInfo {
-        WordInfo::new(text, start, end, 1.0, None)
-    }
-
-    #[test]
-    fn test_split_pool_routes_items_to_two_pools() {
-        // Exercises the real split underlying `split_triplets` with a synthetic
-        // `Pool<u32>` (no model). 4 items, batch 1 → interactive 3, batch 1.
-        let (pool, batch) = Engine::split_pool(vec![1u32, 2, 3, 4], 1);
-        assert_eq!(pool.total(), 3);
-        assert_eq!(batch.as_ref().map(|b| b.total()), Some(1));
-
-        // batch_pool_size 0 → split disabled, no batch pool.
-        let (pool, batch) = Engine::split_pool(vec![1u32, 2, 3, 4], 0);
-        assert_eq!(pool.total(), 4);
-        assert!(batch.is_none());
-
-        // Over-request clamps so at least one triplet stays interactive.
-        let (pool, batch) = Engine::split_pool(vec![1u32, 2, 3], 9);
-        assert_eq!(pool.total(), 1);
-        assert_eq!(batch.as_ref().map(|b| b.total()), Some(2));
-
-        // A single item can't be split.
-        let (pool, batch) = Engine::split_pool(vec![1u32], 1);
-        assert_eq!(pool.total(), 1);
-        assert!(batch.is_none());
-    }
-
-    #[test]
-    fn test_engine_load_missing_dir() {
-        let result = Engine::load_with_pool_size("/nonexistent/path/for/tests", 1);
-        assert!(matches!(result, Err(GigasttError::ModelLoad { .. })));
-    }
-
-    #[test]
-    fn test_engine_load_empty_dir() {
-        let dir = tempfile::tempdir().unwrap();
-        let result = Engine::load_with_pool_size(dir.path().to_str().unwrap(), 1);
-        assert!(matches!(result, Err(GigasttError::ModelLoad { .. })));
-    }
-
-    #[test]
-    fn test_speech_endpoint_policy_matrix() {
-        // Auto: blank without VAD, or VAD fire.
-        assert!(Engine::speech_endpoint(
-            EndpointMode::Auto,
-            true,
-            false,
-            false
-        ));
-        assert!(Engine::speech_endpoint(
-            EndpointMode::Auto,
-            false,
-            true,
-            true
-        ));
-        assert!(!Engine::speech_endpoint(
-            EndpointMode::Auto,
-            true,
-            false,
-            true
-        )); // blank ignored w/ VAD
-        // Assistant: only VAD.
-        assert!(!Engine::speech_endpoint(
-            EndpointMode::Assistant,
-            true,
-            false,
-            false
-        ));
-        assert!(Engine::speech_endpoint(
-            EndpointMode::Assistant,
-            false,
-            true,
-            true
-        ));
-        // Manual: never auto.
-        assert!(!Engine::speech_endpoint(
-            EndpointMode::Manual,
-            true,
-            true,
-            true
-        ));
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_warmup_runs_silent_inference_on_every_triplet() {
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 2)
-            .expect("engine should load");
-        engine
-            .warmup()
-            .expect("warmup must succeed on a working engine");
-        assert_eq!(
-            engine.pool.available(),
-            engine.pool.total(),
-            "every triplet must be returned to the pool after warmup"
-        );
-    }
-
-    // ---- Model-backed coverage (process_chunk / transcribe / state) --------
-    //
-    // These need the GigaAM model on disk; CI / coverage runs them with
-    // `--include-ignored`. They exercise the real streaming + file-decode
-    // branches (empty input, sub-stride, sub-N_FFT, full decode + slide,
-    // silence/short transcription, finish_stream / flush_state).
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_create_state_initial_fields() {
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let state = engine.create_state(false);
-        assert!(state.audio_buffer.is_empty());
-        assert!(state.assembler.is_empty());
-        assert_eq!(state.window_start_samples, 0);
-        assert_eq!(state.context_samples, 0);
-        assert_eq!(state.pending_samples, 0);
-        assert!(state.resampler.is_none());
-        // No VAD attached on a default engine → no endpointer.
-        assert!(state.vad_endpointer.is_none());
-        // Decoder state seeded to blank.
-        assert_eq!(state.decoder.consecutive_blanks, 0);
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_create_state_diarization_flag_ignored_without_feature() {
-        // Without the `diarization` feature the flag is silently ignored and a
-        // perfectly usable state still comes back.
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let state = engine.create_state(true);
-        assert!(state.audio_buffer.is_empty());
-    }
-
-    // Model-free guard for the 2.11.2 diarization fix: `load_speaker_encoder` must
-    // stay wired to polyvoice's `FbankOnnxExtractor` (rank-3 fbank input) via its
-    // 3-arg constructor, NOT the old rank-2 raw-waveform `OnnxEmbeddingExtractor`
-    // (4-arg, with a segment-samples window) that caused the `Got: 2 Expected: 3`
-    // failure. The extractor reads the ONNX model at construction, so a nonexistent
-    // path returns Err (never panics/Ok). This runs on every PR without the ~26 MB
-    // model, and won't compile if the loader's return type / constructor arity
-    // regresses to the waveform extractor.
-    #[cfg(feature = "diarization")]
-    #[test]
-    fn test_load_speaker_encoder_missing_model_errors() {
-        let missing = Path::new("/nonexistent/gigastt-test/wespeaker_resnet34.onnx");
-        let result = diarization::load_speaker_encoder(missing, 1);
-        assert!(
-            result.is_err(),
-            "a missing WeSpeaker model must surface as Err, not panic or Ok"
-        );
-    }
-
-    #[cfg(feature = "diarization")]
-    #[test]
-    #[ignore = "requires the WeSpeaker diarization model"]
-    fn test_speaker_encoder_accepts_waveform_audio() {
-        let model_path =
-            Path::new(&crate::model::default_model_dir()).join("wespeaker_resnet34.onnx");
-        let encoder =
-            diarization::load_speaker_encoder(&model_path, 1).expect("speaker encoder should load");
-        let samples: Vec<f32> = (0..24_000)
-            .map(|i| {
-                let phase = std::f32::consts::TAU * 220.0 * i as f32 / 16_000.0;
-                0.1 * phase.sin()
-            })
-            .collect();
-
-        let embedding = encoder
-            .embed(&samples)
-            .expect("waveform must be converted to rank-3 fbank features");
-
-        assert_eq!(embedding.len(), SPEAKER_EMBEDDING_DIM);
-        assert!(embedding.iter().all(|value| value.is_finite()));
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_process_chunk_empty_input_returns_no_segments() {
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let mut state = engine.create_state(false);
-        let segs = engine
-            .process_chunk(&[], &mut state, &mut guard)
-            .expect("empty chunk must not error");
-        assert!(segs.is_empty(), "empty input yields no segments");
-        assert_eq!(state.audio_buffer.len(), 0);
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_process_chunk_sub_stride_buffers_without_decoding() {
-        // A chunk smaller than the decode stride is buffered and triggers no
-        // decode (the stride gate returns early).
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let mut state = engine.create_state(false);
-        let small = vec![0.0f32; 1600]; // 0.1s ≪ 0.8s stride
-        let segs = engine
-            .process_chunk(&small, &mut state, &mut guard)
-            .expect("sub-stride chunk must not error");
-        assert!(segs.is_empty(), "sub-stride chunk yields no segments yet");
-        assert_eq!(state.audio_buffer.len(), 1600, "samples are buffered");
-        assert_eq!(state.pending_samples, 1600, "pending counter advances");
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_process_chunk_silence_over_stride_decodes_no_words() {
-        // Enough silence to cross the decode stride: the encoder runs but
-        // produces no words (silence), so the partial path returns no segments
-        // and the pending counter resets.
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let mut state = engine.create_state(false);
-        let chunk = vec![0.0f32; 16000]; // 1s of silence, > 0.8s stride
-        let segs = engine
-            .process_chunk(&chunk, &mut state, &mut guard)
-            .expect("decode of silence must not error");
-        // Silence → no words → empty assembler → no partial segment emitted.
-        assert!(segs.is_empty(), "silence decodes to no words");
-        assert_eq!(
-            state.pending_samples, 0,
-            "decode resets the pending counter"
-        );
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_flush_state_empty_returns_none() {
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut state = engine.create_state(false);
-        assert!(
-            engine.flush_state(&mut state).is_none(),
-            "an empty assembler flushes to None"
-        );
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_flush_state_nonempty_returns_final_segment() {
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut state = engine.create_state(false);
-        state.assembler.set_words(vec![word("hello", 0.0, 0.4)]);
-        let seg = engine
-            .flush_state(&mut state)
-            .expect("non-empty assembler flushes to a Final segment");
-        assert!(seg.is_final);
-        assert_eq!(seg.text, "hello");
-        assert!(
-            engine.flush_state(&mut state).is_none(),
-            "finalize resets the assembler"
-        );
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_finish_stream_no_pending_flushes_assembler() {
-        // No buffered audio and no pending samples: finish_stream skips the
-        // forced decode and just flushes whatever the assembler holds.
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let mut state = engine.create_state(false);
-        state.assembler.set_words(vec![word("trailing", 0.0, 0.4)]);
-        let seg = engine
-            .finish_stream(&mut state, &mut guard)
-            .expect("finish_stream flushes the assembler");
-        assert_eq!(seg.text, "trailing");
-        assert!(seg.is_final);
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_finish_stream_empty_state_returns_none() {
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let mut state = engine.create_state(false);
-        assert!(
-            engine.finish_stream(&mut state, &mut guard).is_none(),
-            "an idle stream finishes to None"
-        );
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_transcribe_samples_silence_yields_empty_text() {
-        // The single-pass file path on pure silence: the encoder runs, decode
-        // produces no words, and the result text is empty with a correct
-        // duration.
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let silence = vec![0.0f32; 16000 * 2]; // 2s
-        let result = engine
-            .transcribe_samples(&silence, &mut guard)
-            .expect("silence transcription must not error");
-        assert!(result.text.trim().is_empty(), "silence yields no text");
-        assert!(result.words.is_empty());
-        assert!((result.duration_s - 2.0).abs() < 1e-6);
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_transcribe_samples_short_sub_frame_audio() {
-        // Audio shorter than a single FFT frame: the mel extractor pads to one
-        // zero frame and the encoder still runs — exercising the short-input
-        // single-pass branch without panicking. The decode output is whatever
-        // the model emits on a lone padded frame; we only assert it doesn't
-        // error and the reported duration matches the (tiny) input.
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let tiny = vec![0.0f32; 100]; // < N_FFT (320)
-        let result = engine
-            .transcribe_samples(&tiny, &mut guard)
-            .expect("sub-frame audio must not error");
-        assert!((result.duration_s - 100.0 / 16000.0).abs() < 1e-9);
-    }
-
-    #[test]
-    #[ignore = "requires model"]
-    fn test_transcribe_samples_below_chunk_threshold_single_pass() {
-        // Just under the long-form chunk threshold (30s) takes the single-pass
-        // path; pure silence still yields no words but must not error.
-        let engine = Engine::load_with_pool_size(&crate::model::default_model_dir(), 1)
-            .expect("engine should load");
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let samples = vec![0.0f32; CHUNK_THRESHOLD_SAMPLES]; // exactly at threshold → single pass
-        let result = engine
-            .transcribe_samples(&samples, &mut guard)
-            .expect("at-threshold audio must not error");
-        assert!(result.words.is_empty(), "silence decodes to no words");
-    }
-
-    /// End-to-end proof that the Candle backend transcribes IDENTICALLY to the
-    /// ort backend through the full engine pipeline (mel + encoder + RNN-T
-    /// decode), not just per-stage tensors.
-    ///
-    /// Two engines are built on the SAME model dir — one forced onto the ort CPU
-    /// backend, one forced onto the Candle backend (which reads the sibling
-    /// `candle/*.safetensors`) — and the same fixture wav is transcribed by both.
-    /// Per-stage parity is bit-exact, so the decoded text MUST be byte-identical.
-    ///
-    /// Run with:
-    /// `cargo test -p gigastt-core --features candle --lib -- --ignored --nocapture candle_ort_transcription_parity`
-    #[cfg(feature = "candle")]
-    #[test]
-    #[ignore = "requires v3_rnnt model + candle/*.safetensors"]
-    fn candle_ort_transcription_parity() {
-        let model_dir = crate::model::default_model_dir();
-        let model_path = Path::new(&model_dir);
-
-        let ort_engine = Engine::load_with_factory(
-            model_path,
-            None,
-            1,
-            1,
-            0,
-            Box::new(crate::runtime::ort::factory::OrtFactory::cpu()),
-            1,
-        )
-        .expect("ort engine should load");
-        let candle_engine = Engine::load_with_factory(
-            model_path,
-            None,
-            1,
-            1,
-            0,
-            Box::new(crate::runtime::candle::factory::CandleFactory::new()),
-            1,
-        )
-        .expect("candle engine should load");
-
-        let fixtures = [
-            concat!(
-                env!("CARGO_MANIFEST_DIR"),
-                "/../gigastt/tests/fixtures/golos_00.wav"
-            ),
-            concat!(
-                env!("CARGO_MANIFEST_DIR"),
-                "/../gigastt/tests/fixtures/golos_01.wav"
-            ),
-        ];
-
-        for fixture in fixtures {
-            let mut ort_guard = ort_engine.pool.checkout_blocking().expect("ort checkout");
-            let ort_text = ort_engine
-                .transcribe_file(fixture, &mut ort_guard)
-                .expect("ort transcription")
-                .text;
-
-            let mut candle_guard = candle_engine
-                .pool
-                .checkout_blocking()
-                .expect("candle checkout");
-            let candle_text = candle_engine
-                .transcribe_file(fixture, &mut candle_guard)
-                .expect("candle transcription")
-                .text;
-
-            eprintln!("fixture = {fixture}");
-            eprintln!("ort    = {ort_text:?}");
-            eprintln!("candle = {candle_text:?}");
-
-            assert_eq!(
-                ort_text, candle_text,
-                "candle transcript diverges from ort for {fixture}:\n  ort    = {ort_text:?}\n  candle = {candle_text:?}"
-            );
-        }
-    }
-
-    /// End-to-end proof that the Candle backend produces IDENTICAL output to ort
-    /// through the STREAMING path (sliding-window `process_chunk` + `finish_stream`),
-    /// not just through the whole-file `transcribe_file` path.
-    ///
-    /// Both engines receive the SAME 8 000-sample (0.5 s) chunks fed in order;
-    /// all returned segment texts are concatenated and compared byte-for-byte.
-    /// Per-stage tensor parity is bit-exact, so the streamed transcripts must match.
-    ///
-    /// Run with:
-    /// `cargo test -p gigastt-core --features candle --lib -- --ignored --nocapture candle_ort_streaming_parity`
-    #[cfg(feature = "candle")]
-    #[test]
-    #[ignore = "requires v3_rnnt model + candle/*.safetensors"]
-    fn candle_ort_streaming_parity() {
-        const CHUNK_SAMPLES: usize = 8_000; // 0.5 s at 16 kHz
-
-        let model_dir = crate::model::default_model_dir();
-        let model_path = Path::new(&model_dir);
-
-        let ort_engine = Engine::load_with_factory(
-            model_path,
-            None,
-            1,
-            1,
-            0,
-            Box::new(crate::runtime::ort::factory::OrtFactory::cpu()),
-            1,
-        )
-        .expect("ort engine should load");
-        let candle_engine = Engine::load_with_factory(
-            model_path,
-            None,
-            1,
-            1,
-            0,
-            Box::new(crate::runtime::candle::factory::CandleFactory::new()),
-            1,
-        )
-        .expect("candle engine should load");
-
-        let fixtures = [
-            concat!(
-                env!("CARGO_MANIFEST_DIR"),
-                "/../gigastt/tests/fixtures/golos_00.wav"
-            ),
-            concat!(
-                env!("CARGO_MANIFEST_DIR"),
-                "/../gigastt/tests/fixtures/golos_01.wav"
-            ),
-        ];
-
-        for fixture in fixtures {
-            let samples = audio::decode_audio_file(fixture)
-                .unwrap_or_else(|e| panic!("failed to decode {fixture}: {e:#}"));
-
-            // --- ort streaming ---
-            let mut ort_guard = ort_engine.pool.checkout_blocking().expect("ort checkout");
-            let mut ort_state = ort_engine.create_state(false);
-            let mut ort_text = String::new();
-            for chunk in samples.chunks(CHUNK_SAMPLES) {
-                let segs = ort_engine
-                    .process_chunk(chunk, &mut ort_state, &mut ort_guard)
-                    .expect("ort process_chunk must not error");
-                for seg in segs {
-                    if !ort_text.is_empty() {
-                        ort_text.push(' ');
-                    }
-                    ort_text.push_str(seg.text.trim());
-                }
-            }
-            if let Some(seg) = ort_engine.finish_stream(&mut ort_state, &mut ort_guard) {
-                if !ort_text.is_empty() {
-                    ort_text.push(' ');
-                }
-                ort_text.push_str(seg.text.trim());
-            }
-
-            // --- candle streaming ---
-            let mut candle_guard = candle_engine
-                .pool
-                .checkout_blocking()
-                .expect("candle checkout");
-            let mut candle_state = candle_engine.create_state(false);
-            let mut candle_text = String::new();
-            for chunk in samples.chunks(CHUNK_SAMPLES) {
-                let segs = candle_engine
-                    .process_chunk(chunk, &mut candle_state, &mut candle_guard)
-                    .expect("candle process_chunk must not error");
-                for seg in segs {
-                    if !candle_text.is_empty() {
-                        candle_text.push(' ');
-                    }
-                    candle_text.push_str(seg.text.trim());
-                }
-            }
-            if let Some(seg) = candle_engine.finish_stream(&mut candle_state, &mut candle_guard) {
-                if !candle_text.is_empty() {
-                    candle_text.push(' ');
-                }
-                candle_text.push_str(seg.text.trim());
-            }
-
-            eprintln!("fixture = {fixture}");
-            eprintln!("ort    (streamed) = {ort_text:?}");
-            eprintln!("candle (streamed) = {candle_text:?}");
-
-            assert_eq!(
-                ort_text, candle_text,
-                "candle streamed transcript diverges from ort for {fixture}:\n  ort    = {ort_text:?}\n  candle = {candle_text:?}"
-            );
-        }
-    }
-
-    /// Word-level edit distance / WER between a reference and a hypothesis
-    /// transcript (Levenshtein over whitespace tokens, normalized by reference
-    /// word count). Used by the ANE measurement harness below.
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    fn word_error_rate(reference: &str, hypothesis: &str) -> f64 {
-        let r: Vec<&str> = reference.split_whitespace().collect();
-        let h: Vec<&str> = hypothesis.split_whitespace().collect();
-        if r.is_empty() {
-            return if h.is_empty() { 0.0 } else { 1.0 };
-        }
-        let mut prev: Vec<usize> = (0..=h.len()).collect();
-        let mut cur = vec![0usize; h.len() + 1];
-        for (i, rw) in r.iter().enumerate() {
-            cur[0] = i + 1;
-            for (j, hw) in h.iter().enumerate() {
-                let cost = if rw == hw { 0 } else { 1 };
-                cur[j + 1] = (prev[j] + cost).min(prev[j + 1] + 1).min(cur[j] + 1);
-            }
-            std::mem::swap(&mut prev, &mut cur);
-        }
-        prev[h.len()] as f64 / r.len() as f64
-    }
-
-    /// The 15 Golos fixtures with ground-truth references (from
-    /// `crates/gigastt/tests/fixtures/manifest.json`). `(path, reference)`.
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    fn golos_fixtures() -> Vec<(String, &'static str)> {
-        const REFS: &[(&str, &str)] = &[
-            (
-                "golos_00.wav",
-                "шестьдесят тысяч тенге сколько будет стоить",
-            ),
-            (
-                "golos_01.wav",
-                "покажи мне на смотрешке телеканал синергия тв",
-            ),
-            ("golos_02.wav", "заказать яблоки зеленые"),
-            (
-                "golos_03.wav",
-                "алиса закажи килограммовый торт графские развалины",
-            ),
-            ("golos_04.wav", "ищи телеканал про бизнес на тиви"),
-            ("golos_05.wav", "михаила мурадяна"),
-            (
-                "golos_06.wav",
-                "любовницы две тысячи тринадцать пятнадцатый сезон",
-            ),
-            ("golos_07.wav", "найди боевики"),
-            ("golos_08.wav", "гетто сезон три"),
-            ("golos_09.wav", "хочу посмотреть ростов папа на телевизоре"),
-            ("golos_10.wav", "сбер какое твое самое ненавистное занятие"),
-            ("golos_11.wav", "афина чем платят у китайцев"),
-            (
-                "golos_12.wav",
-                "джой как работает досрочное погашение кредита",
-            ),
-            ("golos_13.wav", "у тебя найдется люк кейдж"),
-            ("golos_14.wav", "у тебя будет лучшая часть пинк"),
-        ];
-        let dir = concat!(env!("CARGO_MANIFEST_DIR"), "/../gigastt/tests/fixtures/");
-        REFS.iter()
-            .map(|(f, r)| (format!("{dir}{f}"), *r))
-            .collect()
-    }
-
-    /// Build the two engines (ort baseline + composite ANE) on the rnnt model in
-    /// the default model dir. Returns `None` (with a printed SKIP) when the
-    /// rnnt model or the bucket-768 `.mlpackage` is absent so the `#[ignore]`d
-    /// measurement tests degrade cleanly on machines without the assets.
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    fn ane_measurement_engines() -> Option<(Engine, Engine)> {
-        let model_dir = crate::model::default_model_dir();
-        let model_path = Path::new(&model_dir);
-
-        let ane_dir = model_path.join("ane");
-        let bucket_768 = ane_dir.join(crate::model::ane_package_dir_name(768));
-        if !crate::model::ane_package_complete(&bucket_768) {
-            eprintln!(
-                "SKIP: ANE bucket-768 package missing in {} (run `gigastt download --ane` or convert locally)",
-                ane_dir.display()
-            );
-            return None;
-        }
-        if ModelVariant::detect_in_dir(model_path).is_none() {
-            eprintln!("SKIP: no model files in {model_dir} (run `gigastt download`)");
-            return None;
-        }
-
-        let ort_engine = Engine::load_with_factory(
-            model_path,
-            None,
-            1,
-            1,
-            0,
-            Box::new(crate::runtime::ort::factory::OrtFactory::cpu()),
-            1,
-        )
-        .expect("ort engine should load");
-        let ane_engine = Engine::load_with_factory(
-            model_path,
-            None,
-            1,
-            1,
-            0,
-            Box::new(crate::runtime::coreml::factory::AneFactory::new()),
-            1,
-        )
-        .expect("ANE engine should load");
-        Some((ort_engine, ane_engine))
-    }
-
-    /// Run one encoder pass directly through a checked-out triplet (mirrors
-    /// [`Engine::run_inference`]'s encoder setup) and return the emitted
-    /// `encoded_len`. Used to compare the ANE and ort encoders' length tensors
-    /// for the SAME mel input, pinning [`calc_output_length`] against ONNX drift.
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    fn encoder_emitted_len(engine: &Engine, features: &[f32], num_frames: usize) -> usize {
-        let mut guard = engine.pool.checkout_blocking().expect("checkout");
-        let triplet = &mut *guard;
-        triplet.encoder_inputs[0].resize_to(Shape::new(vec![1, N_MELS, num_frames]));
-        triplet.encoder_inputs[0]
-            .as_f32_mut()
-            .expect("encoder signal tensor is f32")
-            .copy_from_slice(features);
-        triplet.encoder_inputs[1]
-            .as_i64_mut()
-            .expect("encoder length tensor is i64")[0] = num_frames as i64;
-        let outputs = triplet
-            .encoder
-            .run(&triplet.encoder_inputs)
-            .expect("encoder run");
-        match outputs[1].view().data() {
-            TensorDataView::I32(v) => usize::try_from(v[0]).expect("non-negative len"),
-            TensorDataView::I64(v) => usize::try_from(v[0]).expect("non-negative len"),
-            _ => panic!("unexpected encoder length tensor type"),
-        }
-    }
-
-    /// FULL-GOLOS WER + frame-count-equality measurement (Part 2a + frame pin).
-    ///
-    /// For every Golos fixture, transcribes through BOTH the composite ANE engine
-    /// and the pure-ort baseline, records mel length `T`, the bucket fill % and
-    /// which PATH the ANE encoder took (ANE bucket vs ort fallback), and emits a
-    /// per-clip table plus aggregate WER(ANE vs ort), WER(ANE vs truth) and
-    /// WER(ort vs truth). Additionally asserts the ANE and ort encoders emit the
-    /// SAME `encoded_len` for the same mel input across all fixtures' real `T`
-    /// (pins [`calc_output_length`] == the ort ONNX length op against drift).
-    ///
-    /// ANE parity is SOFT (mask-free FP16 pad-up is not byte-exact: cosine >= 0.94
-    /// at >= 50% fill, a borderline token can flip), so the per-clip ANE-vs-ort
-    /// gate is a small WER threshold rather than byte equality.
-    ///
-    /// Run with:
-    /// `cargo test -p gigastt-core --features ane --lib -- --ignored --nocapture ane_ort_transcription_parity`
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    #[test]
-    #[ignore = "requires v3_rnnt model + ~/.gigastt/models/ane/*.mlpackage + ANE hardware"]
-    fn ane_ort_transcription_parity() {
-        let Some((ort_engine, ane_engine)) = ane_measurement_engines() else {
-            return;
-        };
-
-        // Mirror the encoder-session selection policy (select_bucket over the
-        // shipped ladder) so the table can label the bucket the ANE engine took
-        // without instrumenting the session itself.
-        use crate::model::ANE_BUCKETS;
-        use crate::runtime::coreml::encoder_session::select_bucket;
-        const FILL_FLOOR: f64 = 0.5;
-        // Aggregate gate: the mean WER(ANE vs ort) across all 15 clips must stay
-        // small. Per-clip gate: at most ONE word may differ from ort — the
-        // documented FP16-pad-up borderline-token flip (see FILL_FLOOR) is allowed
-        // on a single word, but a multi-word divergence is a real regression.
-        const MAX_MEAN_WER: f64 = 0.05;
-        const MAX_WORD_DIFF_PER_CLIP: usize = 1;
-
-        let fixtures = golos_fixtures();
-        let mut sum_wer_ane_ort = 0.0;
-        let mut sum_wer_ane_truth = 0.0;
-        let mut sum_wer_ort_truth = 0.0;
-        let mut frame_eq_checked: Vec<usize> = Vec::new();
-        // (clip, word-diff vs ort) for the post-table per-clip gate.
-        let mut clip_word_diffs: Vec<(String, usize)> = Vec::new();
-
-        eprintln!(
-            "\n{:<12} {:>5} {:>6} {:>9} {:>6} texts",
-            "clip", "T", "fill%", "path", "ident"
-        );
-        for (path, reference) in &fixtures {
-            let samples = audio::decode_audio_file(path).expect("decode fixture");
-            let (features, num_frames) = ane_engine.features.compute(&samples);
-            let bucket = select_bucket(num_frames, ANE_BUCKETS, FILL_FLOOR);
-            let on_ane = bucket.is_some();
-            let fill = match bucket {
-                Some(n) => num_frames as f64 / n as f64,
-                None => 0.0,
-            };
-            let path_label = match bucket {
-                Some(n) => format!("ANE-{n}"),
-                None => "ort-fb".to_string(),
-            };
-
-            // Frame-count equality: the ANE and ort encoders must emit the SAME
-            // encoded_len for the same mel input. (On the ort-fallback path this
-            // is trivially the same session class, but we still assert it; on the
-            // ANE path it pins calc_output_length against the ONNX length op.)
-            let ane_len = encoder_emitted_len(&ane_engine, &features, num_frames);
-            let ort_len = encoder_emitted_len(&ort_engine, &features, num_frames);
-            assert_eq!(
-                ane_len, ort_len,
-                "encoded_len mismatch ANE={ane_len} ort={ort_len} for T={num_frames} ({path})"
-            );
-            if on_ane {
-                let formula =
-                    crate::runtime::coreml::encoder_session::calc_output_length(num_frames);
-                assert_eq!(
-                    formula, ort_len,
-                    "calc_output_length({num_frames})={formula} != ort encoder emitted {ort_len}"
-                );
-                frame_eq_checked.push(num_frames);
-            }
-
-            let mut ort_guard = ort_engine.pool.checkout_blocking().expect("ort checkout");
-            let ort_text = ort_engine
-                .transcribe_file(path, &mut ort_guard)
-                .expect("ort transcription")
-                .text;
-            drop(ort_guard);
-
-            let mut ane_guard = ane_engine.pool.checkout_blocking().expect("ANE checkout");
-            let ane_text = ane_engine
-                .transcribe_file(path, &mut ane_guard)
-                .expect("ANE transcription")
-                .text;
-            drop(ane_guard);
-
-            let wer_ane_ort = word_error_rate(&ort_text, &ane_text);
-            let wer_ane_truth = word_error_rate(reference, &ane_text);
-            let wer_ort_truth = word_error_rate(reference, &ort_text);
-            sum_wer_ane_ort += wer_ane_ort;
-            sum_wer_ane_truth += wer_ane_truth;
-            sum_wer_ort_truth += wer_ort_truth;
-
-            // Absolute word-edit distance vs ort (WER * ort word count, rounded).
-            let ort_words = ort_text.split_whitespace().count().max(1);
-            let word_diff = (wer_ane_ort * ort_words as f64).round() as usize;
-
-            let clip = std::path::Path::new(path)
-                .file_stem()
-                .and_then(|s| s.to_str())
-                .unwrap_or(path);
-            clip_word_diffs.push((clip.to_string(), word_diff));
-            eprintln!(
-                "{:<12} {:>5} {:>5.1}% {:>9} {:>6} ort={:?}",
-                clip,
-                num_frames,
-                fill * 100.0,
-                path_label,
-                ort_text == ane_text,
-                ort_text
-            );
-            eprintln!(
-                "{:<12} {:>5} {:>6} {:>9} {:>6} ane={:?}  truth={:?}",
-                "", "", "", "", "", ane_text, reference
-            );
-            eprintln!(
-                "{:<12} WER ANE-vs-ort={wer_ane_ort:.4} (word_diff={word_diff}) ANE-vs-truth={wer_ane_truth:.4} ort-vs-truth={wer_ort_truth:.4}",
-                ""
-            );
-        }
-
-        let n = fixtures.len() as f64;
-        let mean_ane_ort = sum_wer_ane_ort / n;
-        eprintln!("\n=== AGGREGATE (n={}) ===", fixtures.len());
-        eprintln!("mean WER(ANE vs ort)   = {mean_ane_ort:.4}");
-        eprintln!("mean WER(ANE vs truth) = {:.4}", sum_wer_ane_truth / n);
-        eprintln!("mean WER(ort vs truth) = {:.4}", sum_wer_ort_truth / n);
-        eprintln!(
-            "frame-count equality (ANE==ort encoded_len & ==calc_output_length) verified for T = {:?}",
-            frame_eq_checked
-        );
-
-        // Gate AFTER measuring every clip (a measurement harness must not abort
-        // mid-run). Aggregate parity must be tight; per clip, at most one word may
-        // differ from ort (the documented single FP16-pad-up borderline flip).
-        assert!(
-            mean_ane_ort <= MAX_MEAN_WER,
-            "mean WER(ANE vs ort) {mean_ane_ort:.4} > {MAX_MEAN_WER}"
-        );
-        for (clip, diff) in &clip_word_diffs {
-            assert!(
-                *diff <= MAX_WORD_DIFF_PER_CLIP,
-                "clip {clip}: {diff} words differ from ort (> {MAX_WORD_DIFF_PER_CLIP}) — multi-word ANE divergence is a regression, not a borderline flip"
-            );
-        }
-    }
-
-    /// END-TO-END RTFx measurement (Part 2b).
-    ///
-    /// For the fixtures that take the ANE path (>= 384 mel frames, >= 50% fill),
-    /// measures FULL-PIPELINE wall time (audio decode -> mel -> encoder -> RNN-T
-    /// greedy decode -> text) through the ANE engine and the ort baseline, warm
-    /// (first run discarded), median of >= 5. Reports RTFx (audio_secs / median_s)
-    /// for each engine and the speedup ratio ANE/ort. This quantifies how little
-    /// of the encoder-only ~230x ANE speedup survives the CPU-bound RNN-T decode
-    /// loop: the encoder is nearly free on the ANE, but end-to-end the pipeline is
-    /// decode-bound, so the realized full-pipeline speedup is only ~3.7x.
-    ///
-    /// Run with:
-    /// `cargo test -p gigastt-core --features ane --lib -- --ignored --nocapture ane_e2e_rtfx`
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    #[test]
-    #[ignore = "requires v3_rnnt model + ~/.gigastt/models/ane/*.mlpackage + ANE hardware"]
-    fn ane_e2e_rtfx() {
-        let Some((ort_engine, ane_engine)) = ane_measurement_engines() else {
-            return;
-        };
-
-        use crate::model::ANE_BUCKETS;
-        use crate::runtime::coreml::encoder_session::select_bucket;
-        const FILL_FLOOR: f64 = 0.5;
-        const WARM: usize = 1;
-        const TIMED: usize = 6;
-
-        fn median_secs(engine: &Engine, path: &str) -> f64 {
-            // Warmup (discarded) + timed full-pipeline runs.
-            for _ in 0..WARM {
-                let mut g = engine.pool.checkout_blocking().expect("checkout");
-                let _ = engine.transcribe_file(path, &mut g).expect("transcribe");
-            }
-            let mut times: Vec<f64> = Vec::with_capacity(TIMED);
-            for _ in 0..TIMED {
-                let mut g = engine.pool.checkout_blocking().expect("checkout");
-                let t = std::time::Instant::now();
-                let _ = engine.transcribe_file(path, &mut g).expect("transcribe");
-                times.push(t.elapsed().as_secs_f64());
-            }
-            times.sort_by(|a, b| a.partial_cmp(b).unwrap());
-            times[times.len() / 2]
-        }
-
-        eprintln!(
-            "\n{:<12} {:>5} {:>6} {:>6} {:>9} {:>9} {:>9} {:>9} {:>8}",
-            "clip",
-            "T",
-            "bucket",
-            "audio_s",
-            "ort_med_s",
-            "ane_med_s",
-            "ort_RTFx",
-            "ane_RTFx",
-            "speedup"
-        );
-        let mut any_ane = false;
-        for (path, _ref) in golos_fixtures() {
-            let samples = audio::decode_audio_file(&path).expect("decode fixture");
-            let (_features, num_frames) = ane_engine.features.compute(&samples);
-            let Some(bucket) = select_bucket(num_frames, ANE_BUCKETS, FILL_FLOOR) else {
-                continue; // only clips that exercise the ANE encoder path
-            };
-            any_ane = true;
-            let audio_s = samples.len() as f64 / 16000.0;
-            let ort_med = median_secs(&ort_engine, &path);
-            let ane_med = median_secs(&ane_engine, &path);
-            let ort_rtfx = audio_s / ort_med;
-            let ane_rtfx = audio_s / ane_med;
-            let speedup = ort_med / ane_med;
-            let clip = std::path::Path::new(&path)
-                .file_stem()
-                .and_then(|s| s.to_str())
-                .unwrap_or(&path)
-                .to_string();
-            eprintln!(
-                "{:<12} {:>5} {:>6} {:>6.2} {:>9.4} {:>9.4} {:>9.1} {:>9.1} {:>7.2}x",
-                clip, num_frames, bucket, audio_s, ort_med, ane_med, ort_rtfx, ane_rtfx, speedup
-            );
-        }
-        assert!(
-            any_ane,
-            "no Golos fixture took the ANE path (>= 256 mel frames at >= 50% fill); cannot measure e2e RTFx"
-        );
-    }
-
-    /// CONCURRENT-PREDICTION test (Part 1 item 2).
-    ///
-    /// Builds ONE `AneEncoderSession` backed by a single shared `Arc<SharedModel>`
-    /// and fires concurrent `run` calls from N >= 4 threads on the SAME model,
-    /// asserting no crash/panic and that every thread gets the SAME output for the
-    /// same input (deterministic). Exercises the `unsafe impl Send/Sync` under
-    /// real `CPUAndNeuralEngine` multi-thread load.
-    ///
-    /// Run with:
-    /// `cargo test -p gigastt-core --features ane --lib -- --ignored --nocapture ane_concurrent_prediction_deterministic`
-    #[cfg(all(feature = "ane", target_os = "macos"))]
-    #[test]
-    #[ignore = "requires ~/.gigastt/models/ane/gigaam_v3_encoder_768.mlpackage + ANE hardware"]
-    fn ane_concurrent_prediction_deterministic() {
-        use crate::runtime::coreml::bridge;
-        use crate::runtime::coreml::encoder_session::{SharedModel, pad_time};
-        use std::sync::Arc;
-
-        let model_dir = crate::model::default_model_dir();
-        let pkg = Path::new(&model_dir)
-            .join("ane")
-            .join(crate::model::ane_package_dir_name(768));
-        if !crate::model::ane_package_complete(&pkg) {
-            eprintln!("SKIP: ANE bucket-768 package missing at {}", pkg.display());
-            return;
-        }
-
-        // Compile + load ONCE; share across threads (the production sharing model).
-        let model = Arc::new(SharedModel(
-            bridge::compile_and_load(&pkg, true).expect("compile_and_load bucket-768"),
-        ));
-
-        // Deterministic-but-non-trivial mel input padded to the 768 window.
-        const T: usize = 600;
-        const N: usize = 768;
-        let mut mel = vec![0.0f32; N_MELS * T];
-        for (i, v) in mel.iter_mut().enumerate() {
-            *v = ((i % 97) as f32 * 0.013 - 0.5).sin();
-        }
-        let padded: Arc<Vec<f32>> = Arc::new(pad_time(&mel, N_MELS, T, N));
-
-        // Single-threaded reference output.
-        let (reference, ref_shape) =
-            bridge::predict_f32(&model.0, "mel", &padded, &[1, N_MELS, N], "encoded")
-                .expect("reference predict");
-
-        const THREADS: usize = 4;
-        const PER_THREAD: usize = 5;
-        let mut handles = Vec::with_capacity(THREADS);
-        for _ in 0..THREADS {
-            let model = Arc::clone(&model);
-            let padded = Arc::clone(&padded);
-            handles.push(std::thread::spawn(move || {
-                let mut outs = Vec::with_capacity(PER_THREAD);
-                for _ in 0..PER_THREAD {
-                    let (out, shape) =
-                        bridge::predict_f32(&model.0, "mel", &padded, &[1, N_MELS, N], "encoded")
-                            .expect("concurrent predict");
-                    outs.push((out, shape));
-                }
-                outs
-            }));
-        }
-
-        let mut total = 0usize;
-        for h in handles {
-            let outs = h.join().expect("thread did not panic");
-            for (out, shape) in outs {
-                assert_eq!(shape, ref_shape, "concurrent output shape diverged");
-                assert_eq!(
-                    out.len(),
-                    reference.len(),
-                    "concurrent output length diverged"
-                );
-                assert!(
-                    out.iter().all(|v| v.is_finite()),
-                    "concurrent output has non-finite values"
-                );
-                // Bit-for-bit determinism: same model + same input -> same output.
-                assert_eq!(
-                    out, reference,
-                    "concurrent prediction diverged from the single-threaded reference"
-                );
-                total += 1;
-            }
-        }
-        eprintln!(
-            "concurrent OK: {THREADS} threads x {PER_THREAD} predicts = {total} runs, all deterministic & finite"
-        );
-    }
-
-    mod mock_runtime_tests {
-        use std::collections::HashMap;
-        use std::sync::Arc;
-
-        use crate::inference::{EndpointMode, EndpointReason, Engine, PRED_HIDDEN, WordInfo};
-        use crate::runtime::mock::{MockFactory, MockSession};
-        use crate::runtime::tensor::{Shape, Tensor, TensorData};
-
-        const ENC_DIM: usize = 768;
-
-        fn tiny_mock_engine() -> (Engine, tempfile::TempDir) {
-            let tmp = tempfile::tempdir().expect("tempdir");
-            let dir = tmp.path();
-
-            // Empty model files are enough for variant detection; the mock
-            // runtime intercepts all session loading before the filesystem is
-            // read for ONNX data. Product path is INT8-only (stem keys the mock).
-            std::fs::write(dir.join("v3_rnnt_encoder_int8.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_rnnt_decoder.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_rnnt_joint.onnx"), b"").unwrap();
-            // vocab: index 0 = "▁hi", index 1 = "<blk>" (blank wins on ties).
-            std::fs::write(dir.join("v3_vocab.txt"), "\u{2581}hi\n<blk>\n").unwrap();
-
-            let mut sessions: HashMap<String, Arc<MockSession>> = HashMap::new();
-            sessions.insert(
-                "v3_rnnt_encoder_int8".into(),
-                Arc::new(MockSession::new(
-                    vec![Shape::new(vec![1, 64, 1]), Shape::new(vec![1])],
-                    vec![
-                        Tensor::new(
-                            Shape::new(vec![1, ENC_DIM, 1]),
-                            TensorData::F32(vec![0.0; ENC_DIM]),
-                        )
-                        .unwrap(),
-                        Tensor::new(Shape::new(vec![1]), TensorData::I64(vec![1])).unwrap(),
-                    ],
-                )),
-            );
-            sessions.insert(
-                "v3_rnnt_decoder".into(),
-                Arc::new(MockSession::new(
-                    vec![
-                        Shape::new(vec![1, 1]),
-                        Shape::new(vec![1, 1, PRED_HIDDEN]),
-                        Shape::new(vec![1, 1, PRED_HIDDEN]),
-                    ],
-                    vec![
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                    ],
-                )),
-            );
-            sessions.insert(
-                "v3_rnnt_joint".into(),
-                Arc::new(MockSession::new(
-                    vec![
-                        Shape::new(vec![1, ENC_DIM, 1]),
-                        Shape::new(vec![1, PRED_HIDDEN, 1]),
-                    ],
-                    vec![
-                        Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![0.0; 2]))
-                            .unwrap(),
-                    ],
-                )),
-            );
-
-            let factory = Box::new(MockFactory::new(sessions));
-            let engine = Engine::load_with_factory(dir, None, 1, 1, 0, factory, 1)
-                .expect("engine should load with mock runtime");
-            (engine, tmp)
-        }
-
-        #[test]
-        fn test_engine_loads_with_mock_runtime() {
-            let _ = tiny_mock_engine();
-        }
-
-        #[test]
-        fn test_engine_mock_runtime_decodes_silence() {
-            let (engine, _tmp) = tiny_mock_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let samples = vec![0.0f32; 100]; // < N_FFT → one padded mel frame
-            let result = engine
-                .transcribe_samples(&samples, &mut guard)
-                .expect("mock decode must not error");
-
-            assert!(result.text.is_empty(), "blank-only decode yields no text");
-            assert!(result.words.is_empty());
-            assert!((result.duration_s - 100.0 / 16000.0).abs() < 1e-9);
-        }
-
-        #[test]
-        fn test_validate_overrides_truth_table() {
-            use crate::inference::{
-                HotwordOverride, MAX_HOTWORD_PHRASE_CHARS, MAX_HOTWORDS_PER_REQUEST, OverrideError,
-                TranscribeOverrides,
-            };
-
-            // The tiny mock engine loads with no VAD and no punctuator attached,
-            // so any knob turned *on* per-request must be rejected, and any knob
-            // turned *off* (or ITN in either direction) must be accepted.
-            let (engine, _tmp) = tiny_mock_engine();
-            assert!(!engine.has_vad(), "mock engine has no VAD");
-            assert!(!engine.has_punctuator(), "mock engine has no punctuator");
-
-            // All absent → OK (byte-unchanged default path).
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides::default()),
-                Ok(())
-            );
-
-            // vad=Some(true) with no VAD → err vad_not_loaded.
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides {
-                    vad: Some(true),
-                    ..Default::default()
-                }),
-                Err(OverrideError::VadNotLoaded)
-            );
-            // vad=Some(false) is always OK (opting out never needs a resource).
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides {
-                    vad: Some(false),
-                    ..Default::default()
-                }),
-                Ok(())
-            );
-
-            // punctuation=Some(true) with no punctuator → err.
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides {
-                    punctuation: Some(true),
-                    ..Default::default()
-                }),
-                Err(OverrideError::PunctuationNotAvailable)
-            );
-            // punctuation=Some(false) is always OK.
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides {
-                    punctuation: Some(false),
-                    ..Default::default()
-                }),
-                Ok(())
-            );
-
-            // itn=Some(true) is always OK (pure code, no model to load).
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides {
-                    itn: Some(true),
-                    ..Default::default()
-                }),
-                Ok(())
-            );
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides {
-                    itn: Some(false),
-                    ..Default::default()
-                }),
-                Ok(())
-            );
-
-            // Hotword DoS limits live on validate_hotwords (separate from
-            // TranscribeOverrides so that type stays Copy/Eq).
-            use crate::inference::HotwordError;
-            assert_eq!(
-                engine.validate_hotwords(&HotwordOverride::new(vec![], None)),
-                Ok(())
-            );
-            assert_eq!(
-                engine.validate_hotwords(&HotwordOverride::new(
-                    vec!["ok".into(), "fine".into()],
-                    Some(3.0),
-                )),
-                Ok(())
-            );
-
-            let at_cap: Vec<String> = (0..MAX_HOTWORDS_PER_REQUEST)
-                .map(|i| format!("w{i}"))
-                .collect();
-            assert_eq!(
-                engine.validate_hotwords(&HotwordOverride::new(at_cap, None)),
-                Ok(())
-            );
-            let over_cap: Vec<String> = (0..=MAX_HOTWORDS_PER_REQUEST)
-                .map(|i| format!("w{i}"))
-                .collect();
-            assert_eq!(
-                engine.validate_hotwords(&HotwordOverride::new(over_cap, None)),
-                Err(HotwordError::TooManyHotwords)
-            );
-
-            let ok_phrase: String = "а".repeat(MAX_HOTWORD_PHRASE_CHARS);
-            assert_eq!(
-                engine.validate_hotwords(&HotwordOverride::new(vec![ok_phrase], None)),
-                Ok(())
-            );
-            let long_phrase: String = "а".repeat(MAX_HOTWORD_PHRASE_CHARS + 1);
-            assert_eq!(
-                engine.validate_hotwords(&HotwordOverride::new(vec![long_phrase], None)),
-                Err(HotwordError::PhraseTooLong)
-            );
-        }
-
-        #[test]
-        fn test_request_hotword_biaser_semantics() {
-            use crate::inference::{HotwordOverride, TranscribeOverrides};
-
-            // Pin the three-way hotword override contract without requiring a
-            // real speech utterance:
-            // - None → use engine boot biaser (build_request_biaser not used)
-            // - Some(empty) → force off (build_request_biaser returns None)
-            // - Some(phrases) → temporary Biaser when representable
-            let (engine, _tmp) = tiny_mock_engine();
-            // Mock vocab is "▁hi" + blank; "hi" encodes, unknown Cyrillic does not.
-            let engine = engine.with_hotwords(&[("hi".into(), 1.0)], 5.0);
-            assert!(engine.has_hotwords(), "boot biaser attached");
-
-            // Force-off: empty phrase list yields no temporary biaser.
-            let off = HotwordOverride::new(vec![], None);
-            assert!(
-                engine.build_request_biaser(&off).is_none(),
-                "empty override forces biasing off"
-            );
-
-            // Representable phrase builds a temporary biaser.
-            let on = HotwordOverride::new(vec!["hi".into()], Some(7.0));
-            let built = engine
-                .build_request_biaser(&on)
-                .expect("representable phrase should compile");
-            assert_eq!(built.phrase_count(), 1);
-
-            // Default boost path (None) still builds when phrases are present.
-            let default_boost = HotwordOverride::new(vec!["hi".into()], None);
-            assert!(engine.build_request_biaser(&default_boost).is_some());
-
-            // Unrepresentable-only phrases → None (decode continues without bias).
-            let junk = HotwordOverride::new(vec!["яяяя".into()], None);
-            assert!(
-                engine.build_request_biaser(&junk).is_none(),
-                "unrepresentable phrases drop the temporary biaser"
-            );
-
-            // hotwords=None is always valid and leaves the boot biaser in place
-            // (has_hotwords stays true; decode path selects self.biaser).
-            assert_eq!(
-                engine.validate_overrides(&TranscribeOverrides::default()),
-                Ok(())
-            );
-            assert!(engine.has_hotwords());
-        }
-
-        #[test]
-        fn test_ctc_head_attaches_a_biaser() {
-            use crate::inference::HotwordOverride;
-            use crate::model::ModelVariant;
-
-            // A glossary used to be inert here: greedy CTC has no continuation
-            // state, so the biaser was refused outright to stop the engine
-            // claiming otherwise. The prefix beam gives it somewhere to act, so
-            // both the boot biaser and the per-request one attach again.
-            let (mut engine, _tmp) = tiny_mock_engine();
-            engine.variant = ModelVariant::MlCtc;
-
-            let engine = engine.with_hotwords(&[("hi".into(), 1.0)], 5.0);
-            assert!(
-                engine.has_hotwords(),
-                "a CTC head biases through the prefix beam"
-            );
-
-            let per_request = HotwordOverride::new(vec!["hi".into()], Some(7.0));
-            assert!(
-                engine.build_request_biaser(&per_request).is_some(),
-                "a per-request glossary applies on a CTC head too"
-            );
-        }
-
-        #[test]
-        fn test_transcribe_samples_with_overrides_vad_off_matches_default() {
-            use crate::inference::TranscribeOverrides;
-
-            // `?vad=false` on a VAD-less engine is a no-op relative to the default
-            // path (both decode the whole buffer), so the output must be identical.
-            let (engine, _tmp) = tiny_mock_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let samples = vec![0.0f32; 100];
-            let baseline = engine
-                .transcribe_samples(&samples, &mut guard)
-                .expect("baseline decode");
-            let with_vad_off = engine
-                .transcribe_samples_with_overrides(
-                    &samples,
-                    &mut guard,
-                    &TranscribeOverrides {
-                        vad: Some(false),
-                        ..Default::default()
-                    },
-                    None,
-                    false,
-                    None,
-                    super::super::DecodeControls::default(),
-                )
-                .expect("vad-off decode");
-            assert_eq!(baseline.text, with_vad_off.text);
-            assert_eq!(baseline.words.len(), with_vad_off.words.len());
-        }
-
-        #[test]
-        fn test_diarization_requested_without_speaker_model_records_notice() {
-            use crate::inference::{DiarizationOutcome, TranscribeRequest, TranscribeSource};
-            use std::sync::{Arc, OnceLock};
-
-            // A `?diarization=true` request against an engine with no speaker
-            // model must record `NoSpeakerModel` in the outcome sink so the
-            // server can surface it, instead of silently returning an
-            // all-empty-speaker transcript. Model-free: the tiny mock engine
-            // never ships a WeSpeaker model, and a build without the
-            // `diarization` feature reports the same outcome via the same sink.
-            let (engine, _tmp) = tiny_mock_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let samples = vec![0.0f32; 100];
-            let sink: Arc<OnceLock<DiarizationOutcome>> = Arc::new(OnceLock::new());
-            let req = TranscribeRequest::new(TranscribeSource::Samples(&samples))
-                .with_diarization(true)
-                .with_diarization_outcome(Some(sink.clone()));
-            engine
-                .transcribe_request(req, &mut guard)
-                .expect("decode should succeed");
-            assert_eq!(
-                sink.get().copied(),
-                Some(DiarizationOutcome::NoSpeakerModel),
-                "diarization requested without a model must be reported, not silent"
-            );
-        }
-
-        #[test]
-        fn test_create_state_postprocess_overrides_default_to_none() {
-            // A fresh streaming state inherits the engine boot policy: both
-            // per-session overrides start unset.
-            let (engine, _tmp) = tiny_mock_engine();
-            let state = engine.create_state(false);
-            assert_eq!(state.punctuation, None);
-            assert_eq!(state.itn, None);
-        }
-
-        #[test]
-        fn test_flush_state_applies_itn_per_engine_default() {
-            // Streaming finals go through the same ITN pass as the file path:
-            // with the engine default on, number-words are digitized at flush.
-            let (engine, _tmp) = tiny_mock_engine();
-            let engine = engine.with_itn(true);
-            let mut state = engine.create_state(false);
-            state.assembler.set_words(vec![
-                super::word("двадцать", 0.0, 0.4),
-                super::word("один", 0.4, 0.8),
-            ]);
-            let seg = engine.flush_state(&mut state).expect("flush");
-            assert_eq!(seg.text, "21");
-            // Word payloads stay raw — only the joined text is rewritten.
-            assert_eq!(seg.words[0].word, "двадцать");
-            assert_eq!(seg.words[1].word, "один");
-        }
-
-        #[test]
-        fn test_flush_state_session_override_disables_itn() {
-            // `Configure{itn:false}` wins over an ITN-on engine boot policy.
-            let (engine, _tmp) = tiny_mock_engine();
-            let engine = engine.with_itn(true);
-            let mut state = engine.create_state(false);
-            state.itn = Some(false);
-            state.assembler.set_words(vec![
-                super::word("двадцать", 0.0, 0.4),
-                super::word("один", 0.4, 0.8),
-            ]);
-            let seg = engine.flush_state(&mut state).expect("flush");
-            assert_eq!(seg.text, "двадцать один");
-        }
-
-        #[test]
-        fn test_flush_state_default_leaves_text_raw() {
-            // Boot defaults (ITN off, no punctuator) reproduce the pre-feature
-            // behaviour byte-for-byte.
-            let (engine, _tmp) = tiny_mock_engine();
-            let mut state = engine.create_state(false);
-            state.assembler.set_words(vec![
-                super::word("двадцать", 0.0, 0.4),
-                super::word("один", 0.4, 0.8),
-            ]);
-            let seg = engine.flush_state(&mut state).expect("flush");
-            assert_eq!(seg.text, "двадцать один");
-        }
-
-        #[test]
-        fn test_flush_state_punctuation_request_without_punctuator_is_noop() {
-            // `Configure{punctuation:true}` on a punct-less server degrades
-            // gracefully: the final is emitted unchanged (streaming never
-            // errors on post-processing, unlike the REST 409).
-            let (engine, _tmp) = tiny_mock_engine();
-            let mut state = engine.create_state(false);
-            state.punctuation = Some(true);
-            state
-                .assembler
-                .set_words(vec![super::word("hello", 0.0, 0.4)]);
-            let seg = engine.flush_state(&mut state).expect("flush");
-            assert_eq!(seg.text, "hello");
-        }
-
-        /// Engine whose scripted joiner emits one "▁hi" token on its first call
-        /// and blanks afterwards, over a 16-frame encoder output: a single
-        /// decode then crosses the decoder's blank-run endpoint threshold
-        /// (15 consecutive blanks), reproducing a decoder endpoint without a
-        /// model. The encoder expects exactly one decode stride of audio
-        /// (12800 samples → 79 mel frames).
-        fn blank_run_engine() -> (Engine, tempfile::TempDir) {
-            let tmp = tempfile::tempdir().expect("tempdir");
-            let dir = tmp.path();
-
-            std::fs::write(dir.join("v3_rnnt_encoder_int8.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_rnnt_decoder.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_rnnt_joint.onnx"), b"").unwrap();
-            // vocab: index 0 = "▁hi", index 1 = "<blk>".
-            std::fs::write(dir.join("v3_vocab.txt"), "\u{2581}hi\n<blk>\n").unwrap();
-
-            // (12800 - N_FFT) / HOP_LENGTH + 1 mel frames for one stride.
-            const MEL_FRAMES: usize = 79;
-            // 16 frames: frame 0 emits the token, then a blank on the same
-            // frame's second joiner call; frames 1..=15 add 15 more blanks —
-            // 16 consecutive blanks, one above the endpoint threshold.
-            const ENC_LEN: usize = 16;
-
-            let mut sessions: HashMap<String, Arc<MockSession>> = HashMap::new();
-            sessions.insert(
-                "v3_rnnt_encoder_int8".into(),
-                Arc::new(MockSession::new(
-                    vec![Shape::new(vec![1, 64, MEL_FRAMES]), Shape::new(vec![1])],
-                    vec![
-                        Tensor::new(
-                            Shape::new(vec![1, ENC_DIM, ENC_LEN]),
-                            TensorData::F32(vec![0.0; ENC_DIM * ENC_LEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(Shape::new(vec![1]), TensorData::I64(vec![ENC_LEN as i64]))
-                            .unwrap(),
-                    ],
-                )),
-            );
-            sessions.insert(
-                "v3_rnnt_decoder".into(),
-                Arc::new(MockSession::new(
-                    vec![
-                        Shape::new(vec![1, 1]),
-                        Shape::new(vec![1, 1, PRED_HIDDEN]),
-                        Shape::new(vec![1, 1, PRED_HIDDEN]),
-                    ],
-                    vec![
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                    ],
-                )),
-            );
-            sessions.insert(
-                "v3_rnnt_joint".into(),
-                Arc::new(
-                    MockSession::new(
-                        vec![
-                            Shape::new(vec![1, ENC_DIM, 1]),
-                            Shape::new(vec![1, PRED_HIDDEN, 1]),
-                        ],
-                        vec![
-                            Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![0.0; 2]))
-                                .unwrap(),
-                        ],
-                    )
-                    .with_script(vec![
-                        // First joiner call: token 0 ("▁hi") wins the argmax.
-                        vec![
-                            Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![2.0, 0.0]))
-                                .unwrap(),
-                        ],
-                        // All later calls: blank (id 1) wins.
-                        vec![
-                            Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![0.0, 2.0]))
-                                .unwrap(),
-                        ],
-                    ]),
-                ),
-            );
-
-            let factory = Box::new(MockFactory::new(sessions));
-            let engine = Engine::load_with_factory(dir, None, 1, 1, 0, factory, 1)
-                .expect("engine should load with mock runtime");
-            (engine, tmp)
-        }
-
-        #[test]
-        fn test_process_chunk_blank_endpoint_finalizes_without_vad() {
-            // No VAD attached: the decoder's blank-run heuristic is the only
-            // endpoint signal and finalizes the segment as before.
-            let (engine, _tmp) = blank_run_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let mut state = engine.create_state(false);
-            let chunk = vec![0.0f32; 12800]; // one decode stride
-            let segs = engine
-                .process_chunk(&chunk, &mut state, &mut guard)
-                .expect("mock decode must not error");
-            assert_eq!(
-                segs.len(),
-                1,
-                "blank-run endpoint must finalize the segment"
-            );
-            assert!(segs[0].is_final);
-            assert!(segs[0].speech_final);
-            assert_eq!(segs[0].endpoint_reason, Some(EndpointReason::Blank));
-            assert_eq!(segs[0].text, "hi");
-        }
-
-        #[test]
-        fn test_process_chunk_blank_endpoint_ignored_with_vad() {
-            // With a VAD endpointer attached the VAD owns endpointing, so the
-            // decoder's blank-run heuristic must NOT finalize the segment —
-            // otherwise a `--vad-min-silence-ms` above ~600ms would be
-            // unreachable (the decoder would cut the segment first).
-            let (engine, _tmp) = blank_run_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let mut state = engine.create_state(false);
-            state.vad_endpointer = Some(crate::vad::VadEndpointer::new(
-                &crate::vad::VadConfig::default(),
-            ));
-            let chunk = vec![0.0f32; 12800];
-            let segs = engine
-                .process_chunk(&chunk, &mut state, &mut guard)
-                .expect("mock decode must not error");
-            assert_eq!(segs.len(), 1, "decoded words still surface as a partial");
-            assert!(
-                !segs[0].is_final,
-                "blank-run must not finalize while a VAD owns endpointing"
-            );
-            assert!(!segs[0].speech_final);
-            assert!(segs[0].endpoint_reason.is_none());
-            assert_eq!(segs[0].text, "hi");
-        }
-
-        /// Mock engine whose encoder input shape matches a full 2.5 s window
-        /// (STREAM_MAX_WINDOW_SAMPLES → 249 mel frames), so we can hit `over_cap`
-        /// in a single `process_chunk` without multi-stride buffering.
-        fn blank_run_engine_window_cap() -> (Engine, tempfile::TempDir) {
-            let tmp = tempfile::tempdir().expect("tempdir");
-            let dir = tmp.path();
-            std::fs::write(dir.join("v3_rnnt_encoder_int8.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_rnnt_decoder.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_rnnt_joint.onnx"), b"").unwrap();
-            std::fs::write(dir.join("v3_vocab.txt"), "\u{2581}hi\n<blk>\n").unwrap();
-
-            // (40000 - N_FFT) / HOP_LENGTH + 1 = 249 mel frames @ 16 kHz.
-            const MEL_FRAMES: usize = 249;
-            const ENC_LEN: usize = 16;
-
-            let mut sessions: HashMap<String, Arc<MockSession>> = HashMap::new();
-            sessions.insert(
-                "v3_rnnt_encoder_int8".into(),
-                Arc::new(MockSession::new(
-                    vec![Shape::new(vec![1, 64, MEL_FRAMES]), Shape::new(vec![1])],
-                    vec![
-                        Tensor::new(
-                            Shape::new(vec![1, ENC_DIM, ENC_LEN]),
-                            TensorData::F32(vec![0.0; ENC_DIM * ENC_LEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(Shape::new(vec![1]), TensorData::I64(vec![ENC_LEN as i64]))
-                            .unwrap(),
-                    ],
-                )),
-            );
-            sessions.insert(
-                "v3_rnnt_decoder".into(),
-                Arc::new(MockSession::new(
-                    vec![
-                        Shape::new(vec![1, 1]),
-                        Shape::new(vec![1, 1, PRED_HIDDEN]),
-                        Shape::new(vec![1, 1, PRED_HIDDEN]),
-                    ],
-                    vec![
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                        Tensor::new(
-                            Shape::new(vec![1, 1, PRED_HIDDEN]),
-                            TensorData::F32(vec![0.0; PRED_HIDDEN]),
-                        )
-                        .unwrap(),
-                    ],
-                )),
-            );
-            sessions.insert(
-                "v3_rnnt_joint".into(),
-                Arc::new(
-                    MockSession::new(
-                        vec![
-                            Shape::new(vec![1, ENC_DIM, 1]),
-                            Shape::new(vec![1, PRED_HIDDEN, 1]),
-                        ],
-                        vec![
-                            Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![0.0; 2]))
-                                .unwrap(),
-                        ],
-                    )
-                    .with_script(vec![
-                        vec![
-                            Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![2.0, 0.0]))
-                                .unwrap(),
-                        ],
-                        vec![
-                            Tensor::new(Shape::new(vec![1, 1, 2]), TensorData::F32(vec![0.0, 2.0]))
-                                .unwrap(),
-                        ],
-                    ]),
-                ),
-            );
-            let factory = Box::new(MockFactory::new(sessions));
-            let engine = Engine::load_with_factory(dir, None, 1, 1, 0, factory, 1)
-                .expect("engine should load with mock runtime");
-            (engine, tmp)
-        }
-
-        #[test]
-        fn test_process_chunk_window_cap_emits_partial_not_final() {
-            // Hitting STREAM_MAX_WINDOW must commit a stable prefix and emit a
-            // non-final partial — never speech_final. Voice assistants treat
-            // every final as a command; cap-as-final was the Irene bug.
-            let (engine, _tmp) = blank_run_engine_window_cap();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let mut state = engine.create_state(false);
-            // Attach a VAD endpointer so the blank-run heuristic does not fire
-            // first; the cap path is what we exercise.
-            state.vad_endpointer = Some(crate::vad::VadEndpointer::new(
-                &crate::vad::VadConfig::default(),
-            ));
-            // One chunk at the window cap (2.5 s @ 16 kHz).
-            let chunk = vec![0.0f32; 16000 * 5 / 2];
-            let segs = engine
-                .process_chunk(&chunk, &mut state, &mut guard)
-                .expect("mock decode must not error");
-            assert_eq!(segs.len(), 1, "cap must still surface decoded text");
-            assert!(
-                !segs[0].is_final,
-                "window cap must not emit is_final (got final text={:?})",
-                segs[0].text
-            );
-            assert!(!segs[0].speech_final);
-            assert!(segs[0].endpoint_reason.is_none());
-            assert_eq!(segs[0].text, "hi");
-            // Committed prefix survives after cap commit + slide.
-            assert!(
-                !state.assembler.is_empty(),
-                "stable prefix must remain after cap commit"
-            );
-        }
-
-        #[test]
-        fn test_process_chunk_assistant_mode_ignores_blank_without_vad() {
-            let (engine, _tmp) = blank_run_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let mut state = engine.create_state(false);
-            state.endpoint_mode = EndpointMode::Assistant;
-            let chunk = vec![0.0f32; 12800];
-            let segs = engine
-                .process_chunk(&chunk, &mut state, &mut guard)
-                .expect("mock decode must not error");
-            assert_eq!(segs.len(), 1);
-            assert!(
-                !segs[0].is_final,
-                "assistant mode must not finalize on blank-run alone"
-            );
-            assert_eq!(segs[0].text, "hi");
-        }
-
-        #[test]
-        fn test_process_chunk_manual_mode_never_auto_finalizes() {
-            let (engine, _tmp) = blank_run_engine();
-            let mut guard = engine.pool.checkout_blocking().expect("checkout");
-            let mut state = engine.create_state(false);
-            state.endpoint_mode = EndpointMode::Manual;
-            // Even with a synthetic VAD fire, Manual must not close.
-            // We only have blank-run here; Manual ignores it too.
-            let chunk = vec![0.0f32; 12800];
-            let segs = engine
-                .process_chunk(&chunk, &mut state, &mut guard)
-                .expect("mock decode must not error");
-            assert_eq!(segs.len(), 1);
-            assert!(!segs[0].is_final);
-        }
-
-        #[test]
-        fn test_flush_state_marks_stop_endpoint_reason() {
-            let (engine, _tmp) = blank_run_engine();
-            let mut state = engine.create_state(false);
-            state
-                .assembler
-                .append(vec![WordInfo::new("bye", 0.0, 0.3, 0.9, None)]);
-            let seg = engine.flush_state(&mut state).expect("flush");
-            assert!(seg.is_final);
-            assert!(seg.speech_final);
-            assert_eq!(seg.endpoint_reason, Some(EndpointReason::Stop));
-            assert_eq!(seg.text, "bye");
-        }
-    }
-}
