+++ crates/gigastt-core/src/runtime/factory.rs
@@ -6,6 +6,14 @@ pub trait RuntimeFactory: Send + Sync + 'static {
 
     /// Returns a CPU-only factory suitable for small auxiliary models.
     fn cpu_fallback(&self) -> Box<dyn RuntimeFactory>;
+
+    /// Whether [`Engine::load`] should refuse files whose SHA-256 does not
+    /// match the pinned table. Production factories leave this `true`. The
+    /// mock factory returns `false` so unit tests can load empty placeholder
+    /// ONNX files.
+    fn verify_on_disk_checksums(&self) -> bool {
+        true
+    }
 }
 
 /// Owns loaded sessions. One runtime per `Engine`.
